build(pm-gantt): split pending-mutations module + rebuild bundle
Splits pendingMutations.tsx into three files so React Refresh can
fast-reload each independently:

  pendingMutationsCore.ts        — pure state machine (no React)
  pendingMutationsContext.ts     — React context + usePendingMutations()
  usePendingMutationsState.ts    — hook that wires the store to React state
  pendingMutations.tsx           — only PendingMutationsProvider (fast-reload-safe)

Plus rebuild specy_road/pm_gantt_static/ to ship the optimistic-UI
client. New asset hashes:
  index-* and ModalFrame-*  contain optimisticOutline + pending registry
  index-*.css               contains the new outline-row-pending classes

Suites: 406 pytest + 101 vitest, 0 lint errors.
