--- specy_road/pm_gantt_static/assets/index-Do_vJjcc.js
@@ -1,4 +1,0 @@
-const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/EditModal-BEGw50si.js","assets/rolldown-runtime-Dw2cE7zH.js","assets/dnd-stack-Bjh9ZQ0p.js","assets/markdown-stack-BGy33aUx.js","assets/vendor-Dlt94R7z.js","assets/ModalFrame-CHD6Jw2d.js","assets/MarkdownWorkspace-D3ycvloL.js","assets/pmDisplayStatus-CVVg5boq.js","assets/usePmGuiHandlers-C3CRxUMr.js","assets/pmGuiContext-BJC_xsft.js","assets/ConstitutionDrawer-D9R5EKXk.js","assets/ModalPersistStatusFooter-qwFsXZFm.js","assets/SettingsDrawer-B4M0PjtN.js","assets/toolbarIcons-BtpHV_T4.js","assets/SharedDocsDrawer-D9Z9e_j7.js","assets/VisionDrawer-D0kArwWP.js","assets/WorkNotesDrawer-DwaUPJ_M.js"])))=>i.map(i=>d[i]);
-import{r as e}from"./rolldown-runtime-Dw2cE7zH.js";import{a as t,c as n,d as r,f as i,i as a,l as o,m as s,n as c,o as l,p as u,r as d,s as f,t as p,u as m}from"./dnd-stack-Bjh9ZQ0p.js";import{t as h}from"./react-vendor-CWFJq4fC.js";import{O as g,S as _,_ as v,a as y,c as b,d as x,f as S,g as ee,n as C,o as w,p as te,r as T,s as E,t as D,v as ne,w as re,y as ie}from"./ModalFrame-CHD6Jw2d.js";import{a as ae,c as O,i as k,n as A,o as j,r as M,s as oe,t as N}from"./pmDisplayStatus-CVVg5boq.js";import{c as P}from"./markdown-stack-BGy33aUx.js";import{a as se,c as ce,d as le,l as ue,n as de,o as fe,s as pe,t as me}from"./toolbarIcons-BtpHV_T4.js";import{n as he}from"./pmGuiContext-BJC_xsft.js";(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var F=e(s(),1),I=h(),ge=28,_e=28;function ve(e,t){return e?C({left:e.left+ge,top:e.top+_e,width:e.width,height:e.height},{minTop:t}):T({minTop:t})}function L(e,t,n){let r=new Set(e),i=new Map;n.forEach((e,t)=>i.set(e,t));let a=new Map;for(let n of e){let e=t[n]?.node_key;e&&a.set(e,n)}let o=new Map,s=new Map;for(let t of e)o.set(t,0),s.set(t,[]);for(let n of e){let e=t[n]?.dependencies??[];for(let t of e){let e=a.get(t);!e||!r.has(e)||e===n||(s.get(e).push(n),o.set(n,(o.get(n)??0)+1))}}let c=(e,t)=>(i.get(e)??0)-(i.get(t)??0),l=e.filter(e=>o.get(e)===0).sort(c),u=[];for(;l.length;){let e=l.shift();u.push(e);for(let t of s.get(e)??[])o.set(t,(o.get(t)??0)-1),o.get(t)===0&&(l.push(t),l.sort(c))}if(u.length<e.length){let t=e.filter(e=>!u.includes(e)).sort(c);return[...u,...t]}return u}function ye(e,t){let n=window.innerWidth,r=window.innerHeight,i=t+8,a=e.length;if(a===0)return{};let o=n-16,s=r-i-8,c=(o-6*(a-1))/a,l={};return e.forEach((e,n)=>{l[e]=C({left:8+n*(c+6),top:i,width:c,height:s},{minTop:t})}),l}var be=`Complete`,xe=`In Progress`;function Se(){try{return!0}catch{return!0}}function R(e){return(e||``).trim().toLowerCase()}function z(e){return R(e)===`blocked`}function Ce(e){return R(e)===`complete`}function B(e){return R(e)===`in progress`}function we(e){let t=new Map;for(let n of e){let e=n.parent_id===void 0||n.parent_id===null||n.parent_id===``?null:String(n.parent_id),r=n.id;t.has(e)||t.set(e,[]),t.get(e).push(r)}return t}function Te(e,t,n){let r=[],i=new Set;function a(e){if(!i.has(e)){i.add(e);for(let t of n.get(e)??[])a(t);r.push(e)}}for(let n of e){let e=t[n]?.parent_id;(e==null||typeof e==`string`&&e.trim()===``)&&a(n)}for(let t of e)i.has(t)||a(t);return r}function V(e,t,n){let r={};for(let i of e){let e=t[i]??`Not Started`;if(Ce(e)){r[i]=be;continue}if(z(e)){r[i]=e;continue}let a=n.get(i);if(!a||a.length===0){r[i]=e;continue}let o=a.some(e=>B(r[e]));if(a.every(e=>R(r[e])===`complete`)){r[i]=be;continue}if(o){r[i]=xe;continue}r[i]=e}return r}function Ee(e,t,n,r){let i={};for(let r of e)i[r]=N(t[r],n?.[r]);if(!(r?.enabled??Se()))return{...i};let a=we(e.map(e=>t[e]).filter(e=>!!e));return V(Te(e,t,a),i,a)}function De(e,t,n,r){let i=r?.enabled??Se(),a={};for(let r of e)a[r]=N(t[r],n?.[r]);if(!i)return{...a};let o=we(e.map(e=>t[e]).filter(e=>!!e)),s=V(Te(e,t,o),a,o),c={...a};for(let n of e){let e=t[n],r=(o.get(n)?.length||0)>0,i=R(a[n]);r&&s[n]===`In Progress`&&i!==`complete`&&i!==`blocked`&&i!==`in progress`&&(c[n]=xe),e?.type===`phase`&&s[n]===`Complete`&&(c[n]=be)}return c}function Oe(e,t){return!e||e.type!==`phase`||R(e.status||``)===`complete`?!1:R(t)===`complete`}function ke(e,t){let n=new Set,r=e;for(;r;){for(let e of r.dependencies??[])n.add(e);let e=r.parent_id;if(!e)break;r=t[e]}return n}function Ae(e,t,n){let r=new Set,i=new Set,a=[e];for(;a.length;){let e=a.shift();if(i.has(e))continue;i.add(e);let o=t[e];if(o)for(let e of ke(o,t)){let i=n[e]??e;!i||!t[i]||(r.add(i),a.push(i))}}return r.delete(e),r}function je(e,t){if(e.length===0)return 0;let n=1/0;for(let r of e){let e=t[r]??0;e<n&&(n=e)}return n===1/0?0:n}function Me(e){let t=new Set(e.issues.map(e=>e.code)),{resolved:n,config:r}=e,i=r?.integration_branch??n.integration_branch,a=r?.remote??n.remote;if(t.has(`missing_config_file`))return{tone:`red`,label:`No git-workflow.yaml`,tooltipIntro:`Red means the tracked contract file roadmap/git-workflow.yaml is missing. Add it from the specy-road init project template (version, integration_branch, remote) so CLI defaults and this status match your team.`};if(t.has(`invalid_config`))return{tone:`red`,label:`Invalid git-workflow.yaml`,tooltipIntro:`Red means the file exists but failed schema or YAML validation. Fix the file to match schemas/git-workflow.schema.json, then reload.`};if(t.has(`not_git_repo`))return{tone:`yellow`,label:`Git not connected`,tooltipIntro:`Yellow means this directory is not a git worktree, so branch names and trunk refs cannot be checked. Open the project from a clone or set SPECY_ROAD_REPO_ROOT to the repository root.`};if(t.has(`integration_ref_missing`))return{tone:`yellow`,label:`Fetch ${a}/${i}`,tooltipIntro:`Yellow means the contract is valid and git is connected, but there is no local ref for the integration branch yet (${a}/${i}). Run: git fetch ${a}, or create/checkout that branch locally.`};let o=n.git_branch_current,s=n.git_head_short;return{tone:`green`,label:o!=null&&o.length>0?`${i} · ${a} · ${o}`:`${i} · ${a}${s?` · ${s}`:``}`,tooltipIntro:`Green means roadmap/git-workflow.yaml is valid, this tree is a git checkout, and the integration branch ref exists locally. CLI defaults and PM features use these values.`}}function Ne(e,t,n){if(t===void 0||t.enabled!==!0)return e;let{skipped_reason:r,sync_state:i,integration_branch:a,remote:o}=t,s=a,c=o,l=e.tone,u=e.label,d=e.tooltipIntro,f=n&&(r===`dirty_working_tree`||r===`not_on_integration_branch`)?` Merged registry in the PM view can still include claims from ${c}/${s} after git fetch (remote overlay).`:``;return r===`dirty_working_tree`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped (dirty tree)`,d=`${d}\n\nAuto fast-forward of ${s} is skipped because the working tree is not clean. Commit, stash, or discard changes to allow periodic merge --ff-only, or pull ${s} manually.${f}`):r===`not_on_integration_branch`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped (branch)`,d=`${d}\n\nAuto fast-forward runs only when the current branch is the integration branch (${s}).${f}`):r===`not_git_repo`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped`,d=`${d}\n\nAuto fast-forward requires a git worktree.`):r===`integration_ref_unavailable`?(l===`green`&&(l=`yellow`),u=`${u} · no ${c}/${s} ref`,d=`${d}\n\nRun git fetch ${c} so refs/remotes/${c}/${s} exists.`):i===`diverged`?(l===`green`&&(l=`yellow`),u=`${u} · diverged`,d=`${d}\n\nLocal ${s} and ${c}/${s} have diverged; fast-forward is not possible. Resolve with merge/rebase, then pull.`):i===`ahead_of_remote`&&(l===`green`&&(l=`yellow`),u=`${u} · ahead of ${c}`,d=`${d}\n\nThis branch is ahead of ${c}/${s}; push or reset before others can fast-forward.`),{tone:l,label:u,tooltipIntro:d}}var H=P();function Pe({gitWorkflow:e,integrationBranchAutoFf:t,registryRemoteOverlay:n=!1}){if(e===void 0)return(0,H.jsx)(`span`,{className:`app-header-git-workflow app-header-git-workflow--loading`,"aria-hidden":`true`,children:`Git —…`});let r=Ne(Me(e),t,n),i=`git-workflow-status-tip`,{resolved:a}=e;return(0,H.jsxs)(`div`,{className:`app-header-git-workflow-tooltip app-header-doc-tooltip`,children:[(0,H.jsx)(`span`,{id:`git-workflow-status-label`,className:`app-header-git-workflow-label app-header-git-workflow-label--${r.tone}`,tabIndex:0,"aria-describedby":i,children:r.label}),(0,H.jsxs)(`div`,{id:i,role:`tooltip`,className:`app-header-doc-tip app-header-doc-tip--wide app-header-doc-tip--git-workflow`,children:[(0,H.jsx)(`p`,{className:`git-workflow-tip-intro`,children:r.tooltipIntro}),(0,H.jsxs)(`div`,{className:`git-workflow-outline-hint`,children:[(0,H.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Outline`}),(0,H.jsxs)(`p`,{className:`git-workflow-outline-hint-text`,children:[`A green left accent on a row means your named git branch matches the`,` `,(0,H.jsx)(`code`,{children:`branch`}),` field for that task in`,` `,(0,H.jsx)(`code`,{children:`roadmap/registry.yaml`}),` in `,(0,H.jsx)(`em`,{children:`this`}),` checkout (not necessarily on other branches).`]})]}),e.issues.length>0?(0,H.jsxs)(`div`,{className:`git-workflow-issues-block`,children:[(0,H.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Details`}),e.issues.map(e=>(0,H.jsxs)(`div`,{className:`git-workflow-issue`,children:[(0,H.jsx)(`strong`,{children:e.message}),(0,H.jsx)(`div`,{className:`git-workflow-issue-detail`,children:e.detail})]},e.code))]}):null,(0,H.jsxs)(`div`,{className:`git-workflow-resolved-block`,children:[(0,H.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Resolved`}),(0,H.jsxs)(`ul`,{className:`git-workflow-resolved-list`,children:[(0,H.jsxs)(`li`,{children:[`Integration branch:`,` `,(0,H.jsx)(`code`,{children:a.integration_branch})]}),(0,H.jsxs)(`li`,{children:[`Remote: `,(0,H.jsx)(`code`,{children:a.remote})]}),(0,H.jsxs)(`li`,{children:[`Current branch:`,` `,a.git_branch_current!=null&&a.git_branch_current.length>0?a.git_branch_current:`(detached or unnamed)`]}),a.git_head_short?(0,H.jsxs)(`li`,{children:[`HEAD: `,(0,H.jsx)(`code`,{children:a.git_head_short})]}):null]})]})]})]})}var Fe=new Set([`github_pr`,`gitlab_mr`]);function Ie(e){if(!e)return{open:!1,merged:!1,rejected:!1};let t=e.kind||``,n=(e.pr_state||``).toLowerCase(),r=!!e.merged;return n===`merged`||r?{open:!1,merged:!0,rejected:!1}:n===`rejected`||n===`closed`?{open:!1,merged:!1,rejected:!0}:n===`open`||Fe.has(t)?{open:!0,merged:!1,rejected:!1}:{open:!1,merged:!1,rejected:!1}}function Le(e,t,n,r,i,a){let o=(t||``).trim().toLowerCase(),s=(n||``).trim().toLowerCase(),c=a?.[e],l=Ie(c);return l.merged?`var(--bar-complete)`:l.rejected?`var(--bar-mr-rejected)`:s===`blocked`||o===`blocked`?`var(--bar-blocked)`:l.open?`var(--bar-mr-pending)`:O(e,t,r,i)?`var(--bar-in-progress)`:o===`complete`||s===`complete`?`var(--bar-complete)`:o===`not started`||!t?`var(--bar-not-started)`:j(t)?`var(--bar-in-progress)`:`var(--bar-not-started)`}function Re(e){let t=e.node?.status,n=Le(e.nodeId,e.displayStatus,t,e.registryByNode,e.gitCheckoutById,e.gitEnrichment);if(e.selected)return{fill:`var(--gantt-bar-selected)`,stroke:`rgba(0,0,0,0.15)`,strokeWidth:1};if(e.depHighlight)return{fill:`var(--gantt-dep-chain)`,stroke:`rgba(0,0,0,0.15)`,strokeWidth:1};if(e.node?.type===`gate`)return{fill:`var(--bar-gate)`,stroke:`rgba(0,0,0,0.12)`,strokeWidth:1};let r=n===`var(--bar-in-progress)`&&O(e.nodeId,e.displayStatus,e.registryByNode,e.gitCheckoutById);return{fill:n,stroke:r?`var(--accent)`:`rgba(0,0,0,0.15)`,strokeWidth:r?2:1}}var U=38,W=52,ze=.82,G=8;function Be(e){let t=Math.max(1,e);return t<=1?ze*W:(t-1)*W+ze*W}function Ve({orderedIds:e,nodesById:t,displayStatusById:n,gitCheckoutById:r,registryByNode:i,gitEnrichment:a,stackHeaderPx:o=52,depthOffset:s=0,depths:c,spans:l={},edges:u,showInheritedEdges:d=!0,selectedId:f,highlightRowIds:p=null,onSelect:m,onChartBackgroundMouseDown:h}){let g=e.length;if(g===0)return null;let _=e=>Math.max(0,(c[e]??0)-s),v=0;for(let t of e){let e=_(t),n=l[t]??1;v=Math.max(v,e+n)}let y=v+2,b=G+y*W+16,x=Math.max(32,o),S=g*U+x,ee=Math.max(16,x-6),C={};e.forEach((e,t)=>{C[e]=t});let w=d?u:u.filter(e=>e.kind!==`inherited`);return(0,H.jsxs)(`svg`,{className:`gantt-svg`,width:b,height:S,role:`img`,"aria-label":`Dependency Gantt`,children:[h?(0,H.jsx)(`rect`,{x:0,y:0,width:b,height:S,fill:`transparent`,style:{pointerEvents:`all`,cursor:`default`},onMouseDown:e=>{e.preventDefault(),h()}}):null,(0,H.jsx)(`text`,{x:G,y:18,className:`axis-title`,pointerEvents:`none`,children:`Dependency step (not calendar time)`}),(0,H.jsx)(`text`,{x:G+280,y:18,className:`gantt-legend`,fontSize:10,fill:`var(--muted)`,pointerEvents:`none`,children:d?`Solid = explicit dep · Dashed = inherited`:`Solid = explicit dependency`}),Array.from({length:y},(e,t)=>(0,H.jsxs)(`g`,{pointerEvents:`none`,children:[(0,H.jsx)(`line`,{x1:G+t*W,y1:x,x2:G+t*W,y2:S,stroke:`var(--border)`,strokeWidth:1}),(0,H.jsx)(`text`,{x:G+t*W+4,y:ee,fontSize:10,fill:`var(--muted)`,children:s+t+1})]},t)),(0,H.jsx)(`line`,{x1:0,y1:x,x2:b,y2:x,stroke:`var(--border)`,strokeWidth:1,pointerEvents:`none`}),e.map((e,t)=>(0,H.jsx)(`line`,{x1:0,y1:x+t*U+U,x2:b,y2:x+t*U+U,stroke:`var(--border)`,strokeOpacity:.5,strokeWidth:1,pointerEvents:`none`},`h-${e}`)),p&&p.size>0?e.map((e,t)=>p.has(e)?(0,H.jsx)(`rect`,{x:0,y:x+t*U,width:b,height:U,fill:`var(--gantt-dep-chain-bg)`,opacity:.35,pointerEvents:`none`},`dep-hi-${e}`):null):null,e.map((e,o)=>{let s=t[e],c=_(e),u=Be(l[e]??1),d=x+o*U,h=G+c*W,g=f===e,{fill:v,stroke:y,strokeWidth:b}=Re({nodeId:e,selected:g,depHighlight:!!p?.has(e)&&!g,displayStatus:n?.[e]??s?.status,node:s,registryByNode:i,gitCheckoutById:r,gitEnrichment:a});return(0,H.jsx)(`rect`,{x:h,y:d+6,width:u,height:U-12,rx:3,fill:v,stroke:y,strokeWidth:b,style:{cursor:`pointer`,pointerEvents:`all`},onMouseDown:e=>e.stopPropagation(),onClick:()=>m(e)},`bar-${e}`)}),w.map(({from:e,to:t,kind:n})=>{let r=C[e],i=C[t];if(r===void 0||i===void 0)return null;let a=_(e),o=_(t),s=Be(l[e]??1),c=G+a*W+s,u=G+o*W,d=x+r*U+U/2,f=x+i*U+U/2,p=(c+u)/2,m=n===`inherited`;return(0,H.jsx)(`path`,{d:`M ${c} ${d} C ${p} ${d}, ${p} ${f}, ${u} ${f}`,fill:`none`,stroke:`var(--accent)`,strokeWidth:m?1:1.25,strokeOpacity:m?.55:.75,strokeDasharray:m?`5 4`:void 0,pointerEvents:`none`},`${e}->${t}-${n??`x`}`)})]})}var He=e(u(),1);function K(e){let t=e?.parent_id;return t==null||t===``?null:t}function Ue(e,t,n){if(t===n)return!0;let r=new Set,i=K(e[t]);for(;i;){if(i===n)return!0;if(r.has(i))return!1;r.add(i),i=K(e[i])}return!1}function We(e,t,n){return e.includes(n)?e.filter(e=>Ue(t,e,n)):[]}var Ge=`Title and planning are read-only while this task is in active development (in progress, open or merged merge request, or this checkout matches the registered branch).`,q=e=>{let t=n(e);return t.length>0?t:f(e)},Ke=5;function qe(e,t,n,r){let i=Oe(e,n),a=r!==t,o=i&&r.trim().toLowerCase()===`complete`&&t.trim().toLowerCase()!==`complete`,s=[];return i&&s.push(`Display: all descendants complete (roadmap status unchanged).`),a&&!o&&s.push(`PM view reflects active registration or Git remote state; the roadmap file may still list another status until it is updated.`),s.length>0?s.join(`
-
-`):void 0}function J(e){let t=e?.parent_id;return t==null||t===``?null:t}function Je(e,t,n){let r=e[n]?.parent_id;for(;r;){if(r===t)return!0;r=e[r]?.parent_id??null}return!1}function Ye(e,t,n){return t?t===e?!0:Je(n,e,t):!1}function Xe(e,t,n,r,i){let a=r.filter(t=>J(i[t])===e);if(!a.includes(t)||!a.includes(n))return null;let o=a.filter(e=>e!==t),s=o.indexOf(n);return s<0?null:[...o.slice(0,s),t,...o.slice(s)]}function Ze(e,t,n,r,i){let a=r.filter(t=>J(i[t])===e);if(!a.includes(t)||!a.includes(n))return null;let o=a.filter(e=>e!==t),s=o.indexOf(n);if(s<0)return null;let c=s+1;return[...o.slice(0,c),t,...o.slice(c)]}function Qe({nodeId:e}){let{setNodeRef:t,isOver:n}=o({id:`into:${e}`});return(0,H.jsx)(`span`,{ref:t,className:n?`into-drop into-drop-active`:`into-drop`,title:`Drop here to make this row the new parent (last child)`,children:`⎘`})}function $e(){let{setNodeRef:e,isOver:t}=o({id:`into:__root__`});return(0,H.jsx)(`th`,{ref:e,colSpan:Ke,className:t?`into-root into-root-active`:`into-root`,title:`Drop here to move to top level (append as last root)`,children:`Top level`})}function et({targetId:e,onInsertClick:t,depEditActive:n,hidden:r}){let{setNodeRef:i,isOver:a}=o({id:`before:${e}`});return(0,H.jsx)(`tr`,{ref:i,className:`outline-gap-tr`,style:r?{opacity:0,pointerEvents:`none`}:void 0,children:(0,H.jsx)(`td`,{className:a?`outline-gap-cell outline-gap-active`:`outline-gap-cell`,colSpan:Ke,title:`Drop here to insert before this row (same parent as this row)`,children:(0,H.jsx)(`button`,{type:`button`,className:`outline-gap-insert-btn`,"aria-label":`Add task above`,disabled:n,onPointerDown:e=>e.stopPropagation(),onClick:n=>{n.stopPropagation(),t(e)},children:`+`})})})}function tt(e){let{variant:t,id:n,node:r,outlineDepth:i,meta:a,statusText:o,statusCellTitle:s,devText:c,depCellText:l,depEditId:u,titleEditing:d,titleDraft:f,onTitleDraftChange:p,onTitleBlur:m,onTitleKeyDown:h,titleInputRef:g,onDepCellClick:_,depCellRef:v,devCellTitle:y,titleEditLocked:b,rowClass:x,gitCheckoutTitle:S,trRef:ee,trStyle:C,dndListeners:w,dndAttributes:te,onRowDoubleClick:T,onTitleClick:E,onTitlePointerDown:D,onTitlePointerMove:ne,onTitlePointerEnd:re,onIdClick:ie,onStatusDevClick:ae}=e,O=u===n,k=t===`dragPreview`;return(0,H.jsxs)(`tr`,{ref:ee,style:C,className:x||void 0,title:S,...k?{}:w??{},...k?{}:te??{},onDoubleClick:T,children:[(0,H.jsxs)(`td`,{className:`outline-id`,onClick:k?void 0:ie,children:[(0,H.jsx)(`span`,{className:`outline-id-text`,children:r.id}),k?null:(0,H.jsx)(Qe,{nodeId:r.id})]}),(0,H.jsxs)(`td`,{className:`outline-title`,title:b?Ge:void 0,onClick:k?void 0:E,onPointerDown:k?void 0:D,onPointerMove:k?void 0:ne,onPointerUp:k?void 0:re,onPointerCancel:k?void 0:re,onPointerLeave:k?void 0:re,style:{paddingLeft:`${i*12}px`},children:[!k&&d?(0,H.jsx)(`input`,{ref:g,className:`outline-title-input`,value:f,onChange:e=>p(e.target.value),onBlur:m,onKeyDown:h,onClick:e=>e.stopPropagation(),onPointerDownCapture:e=>e.stopPropagation(),onDoubleClick:e=>e.stopPropagation()}):(0,H.jsxs)(`div`,{className:`outline-title-row`,children:[(0,H.jsx)(`div`,{className:`outline-title-text-wrap`,children:(0,H.jsx)(`div`,{children:r.title})}),b?(0,H.jsx)(`span`,{className:`outline-pm-lock`,title:Ge,"aria-label":`Read-only`,children:(0,H.jsxs)(`svg`,{className:`outline-pm-lock-icon`,width:`14`,height:`14`,viewBox:`0 0 24 24`,fill:`none`,xmlns:`http://www.w3.org/2000/svg`,"aria-hidden":`true`,focusable:`false`,children:[(0,H.jsx)(`rect`,{x:`3`,y:`11`,width:`18`,height:`11`,rx:`2`,stroke:`currentColor`,strokeWidth:`2`}),(0,H.jsx)(`path`,{d:`M7 11V7a5 5 0 0 1 10 0v4`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`})]})}):null]}),a?(0,H.jsx)(`div`,{className:`outline-meta`,children:a}):null]}),(0,H.jsx)(`td`,{className:`outline-col-status`,title:s,onClick:k?void 0:ae,children:o}),(0,H.jsx)(`td`,{className:`outline-col-dev`,title:y,onClick:k?void 0:ae,children:c}),(0,H.jsx)(`td`,{ref:v,className:O?`outline-col-dep outline-col-dep-active`:`outline-col-dep`,title:`Click to choose which features this item depends on (explicit)`,onClick:k?void 0:e=>{e.preventDefault(),e.stopPropagation(),_()},children:(0,H.jsx)(`span`,{className:`outline-dep-cell-text`,children:l})})]})}function nt({id:e,node:t,outlineDepth:n,selected:r,meta:a,statusText:o,statusCellTitle:s,devText:l,depCellText:u,depEditId:d,isDepCandidate:f,titleEditing:p,titleDraft:m,onTitleDraftChange:h,onTitleBlur:g,onTitleKeyDown:_,titleInputRef:v,onBeginTitleEdit:y,onSelectRow:b,onOpenModal:x,onDepRowBodyClick:S,onDepCellClick:ee,dragDisabled:C,depCellRef:w,isGitCheckoutRow:te,devCellTitle:T,titleEditLocked:E,rowSourceHidden:D}){let{attributes:ne,listeners:re,setNodeRef:ie,transform:ae,transition:O}=c({id:e,disabled:C}),k={transform:i.Transform.toString(ae),transition:O,opacity:+!D,pointerEvents:D?`none`:void 0},A=(0,F.useRef)(null),j=(0,F.useRef)(null),M=(0,F.useRef)(null),oe=(0,F.useRef)(!1),N=()=>{j.current&&=(clearTimeout(j.current),null),M.current=null};return(0,H.jsx)(tt,{id:e,node:t,outlineDepth:n,selected:r,meta:a,statusText:o,statusCellTitle:s,devText:l,depCellText:u,depEditId:d,isDepCandidate:f,titleEditing:p,titleDraft:m,onTitleDraftChange:h,onTitleBlur:g,onTitleKeyDown:_,titleInputRef:v,onBeginTitleEdit:y,onSelectRow:b,onOpenModal:x,onDepRowBodyClick:S,onDepCellClick:ee,dragDisabled:C,depCellRef:w,isGitCheckoutRow:te,devCellTitle:T,titleEditLocked:E,variant:`interactive`,rowClass:[r?`selected`:``,d===e?`dep-edit-row`:``,f?`dep-candidate-row`:``,te?`outline-row-git-current`:``,t.type===`gate`?`outline-row-gate`:``].filter(Boolean).join(` `),gitCheckoutTitle:te?`Named branch matches the branch field in roadmap/registry.yaml for this row in this checkout.`:void 0,trRef:ie,trStyle:k,dndListeners:re,dndAttributes:ne,onRowDoubleClick:e=>{e.preventDefault(),!d&&(p||(N(),A.current&&=(clearTimeout(A.current),null),x()))},onTitleClick:()=>{if(oe.current){oe.current=!1;return}if(d){S();return}A.current&&clearTimeout(A.current),A.current=setTimeout(()=>{A.current=null,b()},220)},onTitlePointerDown:e=>{d||p||E||e.button===0&&(N(),M.current={x:e.clientX,y:e.clientY},j.current=setTimeout(()=>{j.current=null,M.current=null,oe.current=!0,A.current&&=(clearTimeout(A.current),null),r||b(),y()},500))},onTitlePointerMove:e=>{if(!j.current||!M.current)return;let t=e.clientX-M.current.x,n=e.clientY-M.current.y;t*t+n*n>100&&N()},onTitlePointerEnd:()=>{N()},onIdClick:()=>{A.current&&clearTimeout(A.current),A.current=null,b()},onStatusDevClick:()=>{d?S():b()}})}function rt({orderedIds:e,nodesById:n,rowDepths:i,selectedId:o,prHints:s,gitEnrichment:c,dependencyInheritance:u,registryByNode:f,gitBranchCurrent:h,gitUserName:g,depEditId:_,depDraftKeys:y,onToggleDepCandidate:b,onDepCellActivate:x,onApplyDepEdit:S,onCancelDepEdit:ee,onClearDepDraft:C,onSelect:w,onDoubleClick:te,performRoadmapMutation:T,onMutationError:E,onGapInsert:D,displayStatusById:ne,outlineStatusById:O,reorderLocked:A=!1,interactionLocked:j=!1}){let[N,P]=(0,F.useState)(null),[se,ce]=(0,F.useState)(``),le=(0,F.useRef)(null),ue=r(m(l,{activationConstraint:{distance:6}})),de=(0,F.useMemo)(()=>{let e=new Map;for(let t of Object.values(n))t?.node_key&&e.set(t.node_key,t.id);return e},[n]),fe=e=>{let t=c[e];if(t?.kind===`registry`||t?.kind===`remote_tip`)return``;if(t?.kind===`github_pr`||t?.kind===`gitlab_mr`){let e=t.assignees;return[e?.length?`A: ${e.join(`, `)}`:``].filter(Boolean).join(` · `)||t.hint_line||``}return t?.hint_line?String(t.hint_line):s[e]?s[e].replace(/<br>/g,` · `):``},pe=e=>{let t=de.get(e);if(t)return t;for(let t of Object.values(n))if(t?.node_key===e)return t.id},me=e=>{if(!n[e])return`—`;if(_===e){let e=[...y].map(e=>pe(e)).filter(Boolean);return e.length?[...new Set(e)].sort().join(`, `):`—`}let t=u?.[e]?.explicit??[];return t.length?t.join(`, `):`—`},he=(0,F.useCallback)(async()=>{if(!N)return;if(oe(N,f,h)){P(null);return}let e=n[N];if(!e){P(null);return}let t=se.trim();if(t&&t!==e.title)try{await T(`Saving title…`,()=>ie(N,[{key:`title`,value:t}]))}catch(e){console.error(e),E(String(e))}P(null)},[N,se,n,T,E,f,h]),I=(0,F.useCallback)(()=>{P(null)},[]);(0,F.useEffect)(()=>{if(!N)return;let e=le.current;e&&(e.focus(),e.select())},[N]),(0,F.useEffect)(()=>{if(!N)return;let e=N,t=window.setInterval(()=>{if(oe(e,f,h)){P(null);return}let t=n[e];if(!t)return;let r=se.trim();r&&r!==t.title&&T(`Saving title…`,()=>ie(e,[{key:`title`,value:r}])).catch(e=>E(String(e)))},2500);return()=>window.clearInterval(t)},[N,se,n,T,E,f,h]);let ge=(0,F.useCallback)(e=>{if(N){(async()=>{await he(),x(e)})();return}x(e)},[N,he,x]),_e=e=>{if(e.key===`Escape`){e.preventDefault(),I();return}e.key===`Enter`&&(e.preventDefault(),he())},ve=e=>{if(_||j||oe(e,f,h))return;let t=n[e];t&&(P(e),ce(t.title))},L=(0,F.useCallback)(()=>T(`Refreshing roadmap…`,async()=>{}),[T]),ye=async(t,r)=>{if(t===r)return;let i=n[t],a=n[r];if(!i||!a)return;let o=J(a);if(o!==null&&Ye(t,o,n))return;let s=e.filter(e=>J(n[e])===o&&e!==t).indexOf(r);if(!(s<0)){if(J(i)===o){let i=Xe(o,t,r,e,n);if(!i?.length)return;try{await T(`Updating outline…`,()=>re(o,i))}catch(e){console.error(e),E(String(e)),await L()}return}try{await T(`Updating outline…`,()=>v(i.node_key,o,s))}catch(e){console.error(e),E(String(e)),await L()}}},be=async(t,r)=>{if(t===r)return;let i=n[t],a=n[r];if(!i||!a)return;let o=J(a);if(o!==null&&Ye(t,o,n))return;let s=e.filter(e=>J(n[e])===o&&e!==t).indexOf(r);if(s<0)return;let c=s+1;if(J(i)===o){let i=Ze(o,t,r,e,n);if(!i?.length)return;try{await T(`Updating outline…`,()=>re(o,i))}catch(e){console.error(e),E(String(e)),await L()}return}try{await T(`Updating outline…`,()=>v(i.node_key,o,c))}catch(e){console.error(e),E(String(e)),await L()}},xe=(0,F.useRef)(0);(0,F.useEffect)(()=>{let e=e=>{xe.current=e.clientY};return window.addEventListener(`pointermove`,e,{passive:!0}),()=>window.removeEventListener(`pointermove`,e)},[]);let Se=async t=>{if(Ce(null),_)return;let{active:r,over:i}=t;if(!i)return;let a=String(r.id),o=String(i.id),s=n[a];if(!s)return;if(o.startsWith(`into:`)){let t=o.slice(5),r=t===`__root__`?null:t;if(r===a)return;if(r&&Je(n,a,r)){await L();return}if(Ye(a,r,n))return;let i=e.filter(e=>J(n[e])===r&&e!==a).length;try{await T(`Updating outline…`,()=>v(s.node_key,r,i))}catch(e){console.error(e),E(String(e)),await L()}return}let c;if(c=o.startsWith(`before:`)?o.slice(7):o,a===c||!n[c])return;let l=!1;if(!o.startsWith(`before:`)&&i?.rect){let e=i.rect,t=e.top+e.height/2;l=xe.current>t}l?await be(a,c):await ye(a,c)},R=!!_||A||j,[z,Ce]=(0,F.useState)(null),B=(0,F.useMemo)(()=>z?We(e,n,z):[],[z,e,n]),we=B[0]??null,Te=(0,F.useMemo)(()=>new Set(B),[B]),V=e=>{Ce(String(e.active.id))},Ee=()=>{Ce(null)},De=(0,F.useRef)(null),Oe=(0,F.useRef)(null),[ke,Ae]=(0,F.useState)(null);(0,F.useLayoutEffect)(()=>{if(!_){Ae(null);return}let e=De.current;if(!e){Ae(null);return}let t=()=>{let t=e.getBoundingClientRect(),n=Oe.current?.offsetWidth??0,r=t.right+8;n>0&&r+n>window.innerWidth-8&&(r=t.left-8-n),r<8&&(r=8),n>0&&r+n>window.innerWidth-8&&(r=Math.max(8,window.innerWidth-8-n)),Ae({position:`fixed`,top:t.top+t.height/2,left:r,transform:`translateY(-50%)`,zIndex:42,visibility:`visible`,pointerEvents:`auto`})};t();let n=requestAnimationFrame(()=>t()),r=e.closest(`.outline-wrap`);window.addEventListener(`resize`,t),r?.addEventListener(`scroll`,t,{passive:!0});let i=new ResizeObserver(t);return i.observe(e),r&&i.observe(r),()=>{cancelAnimationFrame(n),window.removeEventListener(`resize`,t),r?.removeEventListener(`scroll`,t),i.disconnect()}},[_,e,y]);let je=(e,t)=>{let r=n[e];if(!r)return null;let a=r?.node_key,l=!!_&&!!a&&y.has(a)&&_!==e,u=oe(e,f,h),d=(r?.status)?.trim()||`Not Started`,p=ne==null?d:ne[e]??d,m=O==null?p:O[e]??p,v=O==null?ne==null?r?.status||`—`:p:m,x=qe(r,d,p,m),S=M(u,p,m);return{id:e,node:r,outlineDepth:i[t]??0,selected:o===e,meta:fe(e),statusText:v,statusCellTitle:x,devText:ae(e,f,c,h,g),depCellText:me(e),depEditId:_,isGitCheckoutRow:u,devCellTitle:k(e,f,c,s),titleEditLocked:S,isDepCandidate:l,titleEditing:N===e,titleDraft:se,onTitleDraftChange:ce,onTitleBlur:()=>void he(),onTitleKeyDown:_e,titleInputRef:le,onBeginTitleEdit:()=>ve(e),dragDisabled:R,onSelectRow:()=>w(e),onOpenModal:()=>te(e),onDepRowBodyClick:()=>{if(_){if(e===_){w(e);return}b(e)}},onDepCellClick:()=>ge(e),depCellRef:_===e?De:void 0}};return(0,H.jsxs)(H.Fragment,{children:[_==null?null:(0,He.createPortal)((0,H.jsxs)(`div`,{ref:Oe,className:`dep-edit-toolbar-portal`,role:`toolbar`,"aria-label":`Dependency edit actions`,style:ke??{position:`fixed`,left:-9999,top:0,visibility:`hidden`,pointerEvents:`none`,zIndex:42},children:[(0,H.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn dep-edit-floating-save`,onClick:e=>{e.preventDefault(),e.stopPropagation(),S()},children:`Save`}),(0,H.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn`,onClick:e=>{e.preventDefault(),e.stopPropagation(),ee()},children:`Cancel`}),(0,H.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn`,disabled:y.size===0,title:`Remove all prerequisites from the draft`,onClick:e=>{e.preventDefault(),e.stopPropagation(),C()},children:`Clear`})]}),document.body),(0,H.jsx)(a,{sensors:ue,collisionDetection:q,onDragStart:V,onDragEnd:Se,onDragCancel:Ee,children:(0,H.jsxs)(H.Fragment,{children:[(0,H.jsxs)(`table`,{className:`outline-table${_?` dep-edit-mode`:``}`,children:[(0,H.jsxs)(`thead`,{children:[(0,H.jsx)(`tr`,{children:(0,H.jsx)($e,{})}),(0,H.jsxs)(`tr`,{children:[(0,H.jsx)(`th`,{className:`outline-id`,children:`ID`}),(0,H.jsx)(`th`,{className:`outline-title`,children:`Task`}),(0,H.jsx)(`th`,{className:`outline-col-status`,children:`Status`}),(0,H.jsx)(`th`,{className:`outline-col-dev`,children:`Dev`}),(0,H.jsx)(`th`,{className:`outline-col-dep`,children:`Dependency`})]})]}),(0,H.jsx)(`tbody`,{children:(0,H.jsx)(p,{items:e,strategy:d,children:e.map((e,t)=>{let n=je(e,t);return n?(0,H.jsxs)(F.Fragment,{children:[(0,H.jsx)(et,{targetId:e,depEditActive:!!_,onInsertClick:D,hidden:!!(z&&we&&Te.has(e)&&e!==we)}),(0,H.jsx)(nt,{...n,rowSourceHidden:!!(z&&Te.has(e))})]},e):null})})})]}),(0,H.jsx)(t,{dropAnimation:null,children:z&&B.length>0?(0,H.jsx)(`table`,{className:`outline-table outline-drag-overlay-table`,children:(0,H.jsx)(`tbody`,{children:B.map(t=>{let n=e.indexOf(t);if(n<0)return null;let r=je(t,n);if(!r)return null;let i=[r.selected?`selected`:``,r.depEditId===t?`dep-edit-row`:``,r.isDepCandidate?`dep-candidate-row`:``,r.isGitCheckoutRow?`outline-row-git-current`:``,r.node.type===`gate`?`outline-row-gate`:``].filter(Boolean).join(` `);return(0,H.jsx)(tt,{...r,variant:`dragPreview`,rowClass:i,gitCheckoutTitle:r.isGitCheckoutRow?`Named branch matches the branch field in roadmap/registry.yaml for this row in this checkout.`:void 0,trStyle:{cursor:`grabbing`}},t)})})}):null})]})})]})}var it=`roadmap: update plan`;function at({open:e,onClose:t,status:n,onRefreshStatus:r,onPublish:i,headerMinTop:a,zIndex:o=200}){let[s,c]=(0,F.useState)(it),[l,u]=(0,F.useState)(!1),[d,f]=(0,F.useState)(null);if((0,F.useEffect)(()=>{e&&(f(null),c(it),r())},[e,r]),!e)return null;let p=!!n?.can_publish&&!l,m=!!n?.blocked,h=n?.detail??null,g=async()=>{let e=s.trim();if(!e){f(`Enter a short description for this update.`);return}f(null),u(!0);try{await i(e),t()}catch(e){f(String(e))}finally{u(!1)}};return(0,H.jsxs)(D,{title:`Publish roadmap changes`,titleId:`publish-roadmap-title`,onClose:t,storageKey:`publish-roadmap`,minTop:a,zIndex:o,bodyClassName:`modal-body--publish`,footer:(0,H.jsxs)(`div`,{className:`modal-footer-actions`,children:[(0,H.jsx)(`button`,{type:`button`,className:`publish-modal-btn publish-modal-btn--ghost`,disabled:l,onClick:t,children:`Cancel`}),(0,H.jsx)(`button`,{type:`button`,className:`publish-modal-btn publish-modal-btn--primary`,disabled:!p,onClick:()=>void g(),children:l?`Publishing…`:`Publish`})]}),children:[(0,H.jsx)(`p`,{className:`outline-meta`,children:`Saves your roadmap, planning, and governance files to git and pushes to the remote so the team can see them. Your edits are already saved on this computer; publishing shares them.`}),n?(0,H.jsxs)(`div`,{className:`publish-roadmap-meta`,children:[(0,H.jsxs)(`div`,{children:[(0,H.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Branch`}),` `,(0,H.jsx)(`code`,{children:n.current_branch??`—`})]}),n.upstream?(0,H.jsxs)(`div`,{children:[(0,H.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Upstream`}),` `,(0,H.jsx)(`code`,{children:n.upstream})]}):null]}):null,m&&h?(0,H.jsxs)(`div`,{className:`publish-roadmap-blocked`,role:`alert`,children:[(0,H.jsx)(`p`,{children:h}),n&&n.out_of_scope_paths.length>0?(0,H.jsx)(`ul`,{className:`publish-roadmap-path-list`,children:n.out_of_scope_paths.slice(0,12).map(e=>(0,H.jsx)(`li`,{children:(0,H.jsx)(`code`,{children:e})},e))}):null,(0,H.jsx)(`p`,{className:`outline-meta`,children:`Ask a developer if you are unsure how to fix this.`})]}):null,!m&&n?.scope_dirty&&n.scope_paths.length>0?(0,H.jsxs)(`div`,{className:`publish-roadmap-scope`,children:[(0,H.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Changed files`}),(0,H.jsx)(`ul`,{className:`publish-roadmap-path-list`,children:n.scope_paths.slice(0,20).map(e=>(0,H.jsx)(`li`,{children:(0,H.jsx)(`code`,{children:e})},e))})]}):null,(0,H.jsx)(`label`,{className:`publish-roadmap-label`,htmlFor:`publish-message`,children:`Summary for teammates`}),(0,H.jsx)(`input`,{id:`publish-message`,type:`text`,className:`publish-roadmap-input`,value:s,onChange:e=>c(e.target.value),disabled:l,autoComplete:`off`,spellCheck:!0,maxLength:500}),d?(0,H.jsx)(`p`,{className:`modal-review-error`,role:`alert`,children:d}):null]})}var Y={splitPct:`pmGanttSplitPct`,refreshSec:`pmGanttRefreshSec`,showInheritedDeps:`pmGanttShowInheritedDeps`,highlightDepChain:`pmGanttHighlightDepChain`,themeMode:`pmGanttThemeMode`};function ot(e,t){return t?`${e}:${t}`:e}function st(e,t){try{let n=ot(e,t),r=localStorage.getItem(n);return r===null&&t&&(r=localStorage.getItem(e),r!==null&&localStorage.setItem(n,r)),r}catch{return null}}function ct(e,t,n){try{localStorage.setItem(ot(e,t),n)}catch{}}function lt(e){try{return localStorage.getItem(e)}catch{return null}}function ut(){let e=Promise.resolve(void 0);return{enqueue(t){let n=e.then(()=>t());return e=n.catch(()=>{}),n}}}function dt(){let e=(0,F.useRef)(null);e.current||=ut();let[t,n]=(0,F.useState)(0),[r,i]=(0,F.useState)(null),a=(0,F.useCallback)((t,r)=>e.current.enqueue(async()=>{i(t),n(e=>e+1);try{await r()}finally{n(e=>{let t=e-1;return t===0&&i(null),t})}}),[]);return{busy:t>0,busyLabel:r,runRoadmapAction:a}}async function ft(e,t){let n=!1;for(;;)try{return await e(),`ok`}catch(e){if(!(e instanceof y))throw e;if(e.retryable&&!n){n=!0,await t();continue}return await t(),`conflict`}}var X=`modulepreload`,pt=function(e){return`/`+e},Z={},mt=function(e,t,n){let r=Promise.resolve();if(t&&t.length>0){let e=document.getElementsByTagName(`link`),i=document.querySelector(`meta[property=csp-nonce]`),a=i?.nonce||i?.getAttribute(`nonce`);function o(e){return Promise.all(e.map(e=>Promise.resolve(e).then(e=>({status:`fulfilled`,value:e}),e=>({status:`rejected`,reason:e}))))}r=o(t.map(t=>{if(t=pt(t,n),t in Z)return;Z[t]=!0;let r=t.endsWith(`.css`),i=r?`[rel="stylesheet"]`:``;if(n)for(let n=e.length-1;n>=0;n--){let i=e[n];if(i.href===t&&(!r||i.rel===`stylesheet`))return}else if(document.querySelector(`link[href="${t}"]${i}`))return;let o=document.createElement(`link`);if(o.rel=r?`stylesheet`:X,r||(o.as=`script`),o.crossOrigin=``,o.href=t,a&&o.setAttribute(`nonce`,a),document.head.appendChild(o),r)return new Promise((e,n)=>{o.addEventListener(`load`,e),o.addEventListener(`error`,()=>n(Error(`Unable to preload CSS for ${t}`)))})}))}function i(e){let t=new Event(`vite:preloadError`,{cancelable:!0});if(t.payload=e,window.dispatchEvent(t),!t.defaultPrevented)throw e}return r.then(t=>{for(let e of t||[])e.status===`rejected`&&i(e.reason);return e().catch(i)})},ht=(0,F.lazy)(()=>mt(()=>import(`./EditModal-BEGw50si.js`).then(e=>({default:e.EditModal})),__vite__mapDeps([0,1,2,3,4,5,6,7,8,9]))),gt=(0,F.lazy)(()=>mt(()=>import(`./ConstitutionDrawer-D9R5EKXk.js`).then(e=>({default:e.ConstitutionDrawer})),__vite__mapDeps([10,1,2,3,5,6,11,8,9]))),_t=(0,F.lazy)(()=>mt(()=>import(`./SettingsDrawer-B4M0PjtN.js`).then(e=>({default:e.SettingsDrawer})),__vite__mapDeps([12,1,2,3,5,13]))),vt=(0,F.lazy)(()=>mt(()=>import(`./SharedDocsDrawer-D9Z9e_j7.js`).then(e=>({default:e.SharedDocsDrawer})),__vite__mapDeps([14,1,2,3,5,8,9]))),yt=(0,F.lazy)(()=>mt(()=>import(`./VisionDrawer-D0kArwWP.js`).then(e=>({default:e.VisionDrawer})),__vite__mapDeps([15,1,2,3,5,6,11,8,9]))),bt=(0,F.lazy)(()=>mt(()=>import(`./WorkNotesDrawer-DwaUPJ_M.js`).then(e=>({default:e.WorkNotesDrawer})),__vite__mapDeps([16,1,2,3,5,6,11,8,9])));function xt(){let e=lt(Y.themeMode);return e===`light`||e===`dark`||e===`system`?e:`system`}function St(e){return Object.fromEntries(e.map(e=>[e.id,e]))}function Ct(){return window.prompt(`Title for new feature`)?.trim()||null}function wt(e){let t=e.trim().replace(/[/\\]+$/,``);if(!t)return``;let n=t.split(/[/\\]/).filter(Boolean);return n[n.length-1]??t}function Tt(){let[e,t]=(0,F.useState)(null),[n,r]=(0,F.useState)(``),[i,a]=(0,F.useState)(null),[o,s]=(0,F.useState)(null),[c,l]=(0,F.useState)(null),[u,d]=(0,F.useState)([]),[f,p]=(0,F.useState)(null),[m,h]=(0,F.useState)({}),[v,C]=(0,F.useState)(!1),[T,D]=(0,F.useState)(null),[re,ae]=(0,F.useState)(null),O=(0,F.useRef)({}),k=(0,F.useRef)(null),j=(0,F.useRef)(0),N=(0,F.useRef)({}),P=(0,F.useRef)(null),[I,ge]=(0,F.useState)(0),[_e,be]=(0,F.useState)(!1),[xe,Se]=(0,F.useState)(!1),[R,z]=(0,F.useState)(!1),[Ce,B]=(0,F.useState)(!1),[we,Te]=(0,F.useState)(!1),[V,Oe]=(0,F.useState)(null),[ke,Me]=(0,F.useState)(!1),[Ne,Fe]=(0,F.useState)(null),[Ie,Le]=(0,F.useState)(()=>{let e=lt(Y.refreshSec);if(e!==null){let t=parseInt(e,10);if(!Number.isNaN(t)&&t>=0&&t<=120)return t}return 5}),Re=(0,F.useRef)(null),[U,W]=(0,F.useState)(()=>{let e=lt(Y.splitPct);if(e){let t=Number(e);if(!Number.isNaN(t)&&t>=15&&t<=85)return t}return 42}),[ze,G]=(0,F.useState)(()=>lt(Y.showInheritedDeps)===`1`),[Be,He]=(0,F.useState)(()=>lt(Y.highlightDepChain)!==`0`),[K,Ue]=(0,F.useState)(!1),[We,Ge]=(0,F.useState)(xt);(0,F.useLayoutEffect)(()=>{if(!i)return;let e=st(Y.themeMode,i);(e===`light`||e===`dark`||e===`system`)&&Ge(e);let t=st(Y.refreshSec,i);if(t!==null){let e=parseInt(t,10);!Number.isNaN(e)&&e>=0&&e<=120&&Le(e)}let n=st(Y.splitPct,i);if(n){let e=Number(n);!Number.isNaN(e)&&e>=15&&e<=85&&W(e)}G(st(Y.showInheritedDeps,i)===`1`),He(st(Y.highlightDepChain,i)!==`0`)},[i]);let[q,Ke]=(0,F.useState)(null),[qe,J]=(0,F.useState)(new Set),Je=(0,F.useRef)(null),Ye=(0,F.useRef)(null),Xe=(0,F.useRef)(null),[Ze,Qe]=(0,F.useState)(52),$e=(0,F.useRef)(!1),et=(0,F.useRef)(!1),tt=(0,F.useRef)(new Set),nt=(0,F.useRef)(null);(0,F.useEffect)(()=>{tt.current=qe},[qe]),(0,F.useEffect)(()=>{nt.current=q},[q]),(0,F.useEffect)(()=>{k.current=f},[f]),(0,F.useEffect)(()=>{j.current=I},[I]),(0,F.useLayoutEffect)(()=>{document.documentElement.setAttribute(`data-theme`,We)},[We]),(0,F.useEffect)(()=>{ct(Y.themeMode,i,We)},[We,i]),(0,F.useLayoutEffect)(()=>{let e=P.current;if(!e)return;let t=()=>ge(e.getBoundingClientRect().bottom);t();let n=new ResizeObserver(t);return n.observe(e),()=>n.disconnect()},[]);let it=(0,F.useCallback)(()=>{let e=Ye.current;if(!e)return;let t=e.querySelector(`thead`),n=e.querySelector(`tbody tr.outline-gap-tr`),r=t?.getBoundingClientRect().height??0,i=n?.getBoundingClientRect().height,a=r+(typeof i==`number`&&i>0?i:8);a>0&&Qe(Math.round(a*10)/10)},[]);(0,F.useLayoutEffect)(()=>{if(!e)return;let t=Ye.current;if(!t)return;let n=new ResizeObserver(()=>{it()});n.observe(t);let r=t.querySelector(`thead`),i=t.querySelector(`tbody`);r&&n.observe(r),i&&n.observe(i);let a=requestAnimationFrame(()=>{it()});return()=>{cancelAnimationFrame(a),n.disconnect()}},[e,it]);let ot=(0,F.useCallback)(async()=>{try{let e=await b();Oe({vision:e.vision_needs_completion,constitution:e.constitution_needs_completion})}catch{Oe(null)}},[]),ut=(0,F.useCallback)(async()=>{try{Fe(await x())}catch{Fe(null)}},[]),X=(0,F.useCallback)(async()=>{s(null);try{let[e,n]=await Promise.all([S(),fetch(`/api/repo`).then(e=>e.json())]);t(e),Re.current=e.fingerprint;let i=n;r(i.repo_root||``),a(typeof i.repo_id==`string`?i.repo_id:null),l(t=>t&&e.ordered_ids.includes(t)?t:e.ordered_ids[0]??null);try{Re.current=await te()}catch{}ot(),ut()}catch(e){s(String(e))}},[ot,ut]),pt=(0,F.useCallback)(async e=>{try{await _(e)}catch(e){throw e instanceof y&&await X(),e}await X()},[X]),{busy:Z,busyLabel:mt,runRoadmapAction:Tt}=dt(),Q=(0,F.useCallback)((e,t)=>Tt(e,async()=>{if(await ft(t,X)===`conflict`){s(`Roadmap or workspace changed elsewhere; the view was refreshed. Retry if needed.`);return}await X()}),[Tt,X]),Et=(0,F.useMemo)(()=>({onConcurrencyConflict:X}),[X]);(0,F.useEffect)(()=>{g(()=>Re.current)},[]),(0,F.useEffect)(()=>{X()},[X]),(0,F.useEffect)(()=>{let e=window.setInterval(()=>{ut()},12e3);return()=>window.clearInterval(e)},[ut]),(0,F.useEffect)(()=>{ct(Y.splitPct,i,String(U))},[U,i]),(0,F.useEffect)(()=>{ct(Y.showInheritedDeps,i,ze?`1`:`0`)},[ze,i]),(0,F.useEffect)(()=>{ct(Y.highlightDepChain,i,Be?`1`:`0`)},[Be,i]),(0,F.useEffect)(()=>{ct(Y.refreshSec,i,String(Ie))},[Ie,i]),(0,F.useEffect)(()=>{if(Ie<=0)return;let e=window.setInterval(()=>{(async()=>{try{let e=await te(),t=Re.current;t!==null&&e!==t?await Tt(`Checking for updates…`,()=>X()):Re.current=e}catch{}})()},Ie*1e3);return()=>window.clearInterval(e)},[Ie,Tt,X]);let Dt=(0,F.useCallback)(()=>{Ke(null),J(new Set)},[]),Ot=(0,F.useCallback)(()=>{J(new Set)},[]),kt=(0,F.useCallback)(async()=>{let e=nt.current;if(!e)return;let t=[...tt.current].sort().join(` `);try{s(null),await Q(`Saving dependencies…`,async()=>{await ie(e,[{key:`dependencies`,value:t}]),Dt()})}catch(e){s(String(e))}},[Q,Dt]),At=(0,F.useMemo)(()=>wt(n),[n]),$=(0,F.useMemo)(()=>e?St(e.nodes):{},[e]),jt=(0,F.useMemo)(()=>{if(!c)return!0;let e=$[c]?.parent_id;return e==null||e===``},[c,$]),Mt=(0,F.useMemo)(()=>e?.ordered_ids?De(e.ordered_ids,$,e.registry_by_node):{},[e,$]),Nt=(0,F.useMemo)(()=>e?.ordered_ids?Ee(e.ordered_ids,$,e.registry_by_node):{},[e,$]),Pt=(0,F.useMemo)(()=>e?.ordered_ids?K?e.ordered_ids.filter(e=>(Nt[e]??``).trim().toLowerCase()!==`complete`):e.ordered_ids:[],[e?.ordered_ids,K,Nt]),Ft=(0,F.useMemo)(()=>{if(!e?.ordered_ids||!e.row_depths)return[];if(!K)return e.row_depths;let t=[];for(let n=0;n<e.ordered_ids.length;n++)(Nt[e.ordered_ids[n]]??``).trim().toLowerCase()!==`complete`&&t.push(e.row_depths[n]??0);return t},[e?.ordered_ids,e?.row_depths,K,Nt]),It=(0,F.useMemo)(()=>!K||!e?.dependency_depths?0:je(Pt,e.dependency_depths),[K,e?.dependency_depths,Pt]);(0,F.useEffect)(()=>{if(!K||!e?.ordered_ids?.length)return;let t=e=>(Nt[e]??``).trim().toLowerCase()===`complete`;l(n=>n&&!t(n)?n:e.ordered_ids.find(e=>!t(e))??null)},[K,e?.ordered_ids,Nt]),(0,F.useEffect)(()=>{!K||!q||(Nt[q]??``).trim().toLowerCase()===`complete`&&Dt()},[K,q,Nt,Dt]);let Lt=(0,F.useMemo)(()=>{if(!e?.ordered_ids)return{};let t=e.registry_by_node??{},n=e.git_enrichment??{},r={};for(let i of e.ordered_ids)r[i]=A($[i],t[i],n[i],Mt[i]);return r},[e,$,Mt]),Rt=(0,F.useMemo)(()=>{if(!e?.ordered_ids)return{};let t=e.registry_by_node??{},n=e.git_workflow?.resolved?.git_branch_current??null,r={};for(let i of e.ordered_ids)r[i]=oe(i,t,n);return r},[e]),zt=(0,F.useMemo)(()=>{if(!c)return!1;let e=($[c]?.status)?.trim()||`Not Started`,t=Mt[c]??e,n=Lt[c]??t;return M(!!Rt[c],t,n)},[c,$,Mt,Lt,Rt]),Bt=(0,F.useMemo)(()=>e?.nodes?Object.fromEntries(e.nodes.map(e=>[e.node_key,e.id])):{},[e]),Vt=(0,F.useMemo)(()=>!Be||!c?null:Ae(c,$,Bt),[Be,c,$,Bt]),Ht=(0,F.useCallback)(t=>{let n=e?.nodes.find(e=>e.id===t);n&&(Ke(t),J(new Set(n.dependencies??[])))},[e?.nodes]),Ut=(0,F.useCallback)(e=>{if(q===e){kt();return}Ht(e)},[q,kt,Ht]),Wt=(0,F.useCallback)(e=>{if(!q)return;let t=$[q]?.node_key,n=$[e]?.node_key;!t||!n||n===t||J(e=>{let t=new Set(e);return t.has(n)?t.delete(n):t.add(n),t})},[q,$]),Gt=(0,F.useCallback)((e,t)=>{O.current[e]=t,h(t=>{if(!(e in t))return t;let n={...t};return delete n[e],n})},[]),Kt=(0,F.useCallback)(e=>{d(t=>{let n=!t.includes(e),r=t.includes(e)?[...t.filter(t=>t!==e),e]:[...t,e];if(n){let n=k.current??t[t.length-1]??null,r=ve(n?O.current[n]:void 0,j.current);h(t=>({...t,[e]:r}))}return r}),p(e)},[]),qt=(0,F.useCallback)(e=>{p(e),d(t=>t.includes(e)?[...t.filter(t=>t!==e),e]:t)},[]),Jt=(0,F.useCallback)(e=>{d(t=>t.filter(t=>t!==e)),delete O.current[e],h(t=>{if(!(e in t))return t;let n={...t};return delete n[e],n})},[]),Yt=(0,F.useCallback)(()=>{!e||u.length===0||(v?(ae({...N.current}),C(!1),D(null),requestAnimationFrame(()=>{requestAnimationFrame(()=>ae(null))})):(N.current={...O.current},D(ye(L(u,$,e.ordered_ids),I)),C(!0)))},[e,u,$,v,I]);(0,F.useEffect)(()=>{f!=null&&!u.includes(f)&&p(u.length?u[u.length-1]:null)},[u,f]),(0,F.useEffect)(()=>{u.length===0&&v&&(C(!1),D(null))},[u.length,v]),(0,F.useEffect)(()=>{!v||!e||u.length===0||D(ye(L(u,$,e.ordered_ids),I))},[v,e,u,$,I]);let Xt=!c||e?.outline_actions!=null&&c!=null&&e.outline_actions[c]?.can_indent===!1,Zt=!c||e?.outline_actions!=null&&c!=null&&e.outline_actions[c]?.can_outdent===!1;(0,F.useEffect)(()=>{let e=e=>{if(u.length>0||_e||xe||R)return;let t=e.target;if(t instanceof HTMLElement){if(t.isContentEditable)return;let e=t.tagName;if(e===`INPUT`||e===`TEXTAREA`||e===`SELECT`)return}if(e.key===`Escape`&&q){e.preventDefault(),Dt();return}if(e.key===`Enter`&&q){if(t instanceof HTMLElement){let e=t.tagName;if(e===`BUTTON`||e===`TEXTAREA`||e===`A`)return}e.preventDefault(),kt();return}if(e.key===`Tab`&&c&&!q){if(t instanceof HTMLElement&&t.closest(`.app-header`))return;e.preventDefault(),e.shiftKey?Zt||Q(`Updating outline…`,async()=>{await ne(c)}).catch(e=>s(String(e))):Xt||Q(`Updating outline…`,async()=>{await ee(c)}).catch(e=>s(String(e)))}};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[u.length,_e,xe,R,q,c,Xt,Zt,Q,Dt,kt]);let Qt=e=>{let t=Ye.current,n=Xe.current;!t||!n||$e.current||($e.current=!0,e===`left`?n.scrollTop=t.scrollTop:t.scrollTop=n.scrollTop,requestAnimationFrame(()=>{$e.current=!1}))},$t=e=>{e.preventDefault(),et.current=!0;let t=Je.current;if(!t)return;e.target.setPointerCapture(e.pointerId);let n=e=>{if(!et.current)return;let n=t.getBoundingClientRect(),r=(e.clientX-n.left)/n.width*100;W(Math.min(85,Math.max(15,r)))},r=t=>{et.current=!1,e.target.releasePointerCapture(t.pointerId),window.removeEventListener(`pointermove`,n),window.removeEventListener(`pointerup`,r)};window.addEventListener(`pointermove`,n),window.addEventListener(`pointerup`,r)},en=()=>{if(!c)return;let e=Ct();e&&Q(`Adding task…`,async()=>{await w(c,`above`,e,`task`)}).catch(e=>s(String(e)))},tn=()=>{if(!c)return;let e=Ct();e&&Q(`Adding task…`,async()=>{await w(c,`below`,e,`task`)}).catch(e=>s(String(e)))},nn=()=>{if(!c)return;let e=Ct();e&&Q(`Inserting gate…`,async()=>{await w(c,`above`,e,`gate`)}).catch(e=>s(String(e)))},rn=e=>{let t=Ct();t&&Q(`Adding task…`,async()=>{await w(e,`above`,t,`task`)}).catch(e=>s(String(e)))},an=()=>{if(!c)return;let e=$[c]?.id??c;window.confirm(`Remove task "${e}" from the roadmap? This cannot be undone.`)&&(s(null),Q(`Removing task…`,async()=>{await E(c)}).catch(e=>s(String(e))))},on=!!(Ne?.scope_dirty&&!Ne?.blocked&&!Z);return(0,H.jsx)(he,{value:Et,children:(0,H.jsxs)(`div`,{className:`app-shell`,children:[Z?(0,H.jsxs)(`div`,{className:`roadmap-sync-banner`,role:`status`,"aria-live":`polite`,"aria-label":mt??`Working`,children:[(0,H.jsx)(`span`,{className:`roadmap-sync-spinner`,"aria-hidden":`true`}),(0,H.jsx)(`span`,{children:mt??`Working…`})]}):null,(0,H.jsxs)(`div`,{className:`app-chrome`,"aria-busy":Z,inert:Z?!0:void 0,children:[(0,H.jsxs)(`header`,{ref:P,className:`app-header`,children:[(0,H.jsxs)(`div`,{className:`app-header-row1`,children:[(0,H.jsx)(`div`,{className:`app-header-row1-left`,children:(0,H.jsxs)(`h1`,{className:`app-title-line`,children:[(0,H.jsx)(`span`,{className:`app-title-core`,children:`specy-road — PM Gantt`}),At?(0,H.jsxs)(H.Fragment,{children:[(0,H.jsxs)(`span`,{className:`repo-root-sep`,"aria-hidden":`true`,children:[` `,`-`,` `]}),(0,H.jsx)(`span`,{className:`repo-root-folder`,title:n,"aria-label":`Repository root: ${n}`,children:At})]}):null]})}),(0,H.jsxs)(`div`,{className:`app-header-row1-actions`,children:[(0,H.jsx)(Pe,{gitWorkflow:e?.git_workflow,integrationBranchAutoFf:e?.integration_branch_auto_ff,registryRemoteOverlay:e?.registry_overlay?.enabled===!0}),(0,H.jsx)(`button`,{type:`button`,className:`app-header-icon-btn app-header-tile-btn`,"aria-pressed":K,disabled:Z,title:K?`Show all roadmap rows, including completed work`:`Hide rows whose subtree is effectively complete`,"aria-label":K?`Show completed work`:`Hide completed work`,onClick:()=>Ue(e=>!e),children:K?`Show Complete`:`Hide Complete`}),u.length>0?(0,H.jsx)(`button`,{type:`button`,className:`app-header-icon-btn app-header-tile-btn`,"aria-pressed":v,disabled:Z,title:v?`Restore task dialogs to their positions before tiling`:`Tile open task dialogs by dependency, left to right`,"aria-label":v?`Untile task dialogs`:`Tile task dialogs`,onClick:()=>Yt(),children:`Tile`}):null,(0,H.jsx)(`button`,{type:`button`,className:`app-header-icon-btn`,"aria-label":`Settings`,title:`Settings`,disabled:Z,onClick:()=>be(!0),children:(0,H.jsx)(me,{})})]})]}),(0,H.jsx)(`div`,{className:`app-header-row2`,children:(0,H.jsxs)(`div`,{className:`app-header-row2-inner`,children:[(0,H.jsxs)(`div`,{className:`app-header-toolbar`,children:[(0,H.jsx)(`button`,{type:`button`,className:on?`toolbar-icon-btn toolbar-icon-btn--publish-ready`:`toolbar-icon-btn`,disabled:Z,title:Ne?.blocked?`Publish blocked — see dialog for details`:on?`Publish roadmap changes (ready to share)`:`Publish roadmap changes to the remote repository`,"aria-label":`Publish roadmap changes`,onClick:()=>{Me(!0),ut()},children:(0,H.jsx)(pe,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||!c||zt,title:zt?`Editing is disabled while this task is in active development (in progress, MR state, or checkout matches the registered branch)`:`Edit selected task`,"aria-label":`Edit selected task`,onClick:()=>{c&&Kt(c)},children:(0,H.jsx)(fe,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||Xt,title:`Indent`,"aria-label":`Indent`,onClick:()=>{c&&Q(`Updating outline…`,async()=>{await ee(c)}).catch(e=>s(String(e)))},children:(0,H.jsx)(de,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||Zt,title:`Outdent`,"aria-label":`Outdent`,onClick:()=>{c&&Q(`Updating outline…`,async()=>{await ne(c)}).catch(e=>s(String(e)))},children:(0,H.jsx)(se,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||!c,title:`Add task above selection`,"aria-label":`Add task above selection`,onClick:en,children:(0,H.jsx)(ce,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||!c,title:`Add task below selection`,"aria-label":`Add task below selection`,onClick:tn,children:(0,H.jsx)(ue,{})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:Z||!c||jt,title:c?jt?`Gate rows must sit under a vision or phase — select an indented row`:`Insert gate at selection (above this row); display ids renumber. PM hold, not dev pickup`:`Select a row first`,"aria-label":`Insert gate at selection`,onClick:nn,children:(0,H.jsx)(`span`,{className:`toolbar-gate-icon`,"aria-hidden":`true`,children:`G`})}),(0,H.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn toolbar-icon-btn-danger`,disabled:Z||!c,title:`Delete selected row`,"aria-label":`Delete selected row`,onClick:an,children:(0,H.jsx)(le,{})})]}),(0,H.jsxs)(`div`,{className:`app-header-docs-row`,children:[(0,H.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,H.jsxs)(`div`,{className:V?.vision?`app-header-doc-tooltip app-header-doc-tooltip--incomplete`:`app-header-doc-tooltip`,children:[(0,H.jsx)(`button`,{type:`button`,className:V?.vision?`app-header-doc-btn app-header-doc-btn--incomplete`:`app-header-doc-btn`,"aria-describedby":V?.vision?`gov-tip-vision`:void 0,onClick:()=>z(!0),children:`Vision`}),V?.vision?(0,H.jsxs)(`div`,{id:`gov-tip-vision`,role:`tooltip`,className:`app-header-doc-tip`,children:[`Needs completion — replace the starter text in`,` `,(0,H.jsx)(`code`,{children:`vision.md`}),` with your product vision.`]}):null]})}),(0,H.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,H.jsxs)(`div`,{className:V?.constitution?`app-header-doc-tooltip app-header-doc-tooltip--incomplete`:`app-header-doc-tooltip`,children:[(0,H.jsx)(`button`,{type:`button`,className:V?.constitution?`app-header-doc-btn app-header-doc-btn--incomplete`:`app-header-doc-btn`,"aria-describedby":V?.constitution?`gov-tip-constitution`:void 0,onClick:()=>Se(!0),children:`Constitution`}),V?.constitution?(0,H.jsxs)(`div`,{id:`gov-tip-constitution`,role:`tooltip`,className:`app-header-doc-tip`,children:[`Needs completion — customize`,` `,(0,H.jsx)(`code`,{children:`purpose.md`}),` and `,(0,H.jsx)(`code`,{children:`principles.md`}),` `,`beyond the scaffold templates.`]}):null]})}),(0,H.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,H.jsx)(`button`,{type:`button`,className:`app-header-doc-btn`,onClick:()=>B(!0),children:`Shared docs`})}),(0,H.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,H.jsx)(`button`,{type:`button`,className:`app-header-doc-btn`,onClick:()=>Te(!0),children:`Session notes`})})]})]})})]}),o?(0,H.jsx)(`p`,{style:{padding:`0 0.75rem`,color:`crimson`},children:o}):null,e?(0,H.jsxs)(`div`,{className:`split`,ref:Je,children:[(0,H.jsx)(`div`,{className:`outline-wrap`,ref:Ye,style:{flex:`0 0 ${U}%`,width:`${U}%`},onScroll:()=>Qt(`left`),children:(0,H.jsx)(rt,{orderedIds:Pt,nodesById:$,displayStatusById:Mt,outlineStatusById:Lt,rowDepths:Ft,reorderLocked:K,interactionLocked:Z,selectedId:c,prHints:e.pr_hints,gitEnrichment:e.git_enrichment,dependencyInheritance:e.dependency_inheritance,registryByNode:e.registry_by_node,gitBranchCurrent:e.git_workflow?.resolved?.git_branch_current??null,gitUserName:e.git_workflow?.resolved?.git_user_name??null,depEditId:q,depDraftKeys:qe,onToggleDepCandidate:Wt,onDepCellActivate:Ut,onApplyDepEdit:()=>void kt(),onCancelDepEdit:Dt,onClearDepDraft:Ot,onSelect:l,onDoubleClick:e=>{l(e),Kt(e)},performRoadmapMutation:Q,onMutationError:e=>s(e),onGapInsert:rn})}),(0,H.jsx)(`div`,{className:`split-resize-handle`,role:`separator`,"aria-orientation":`vertical`,"aria-label":`Resize feature list and chart`,onPointerDown:$t}),(0,H.jsx)(`div`,{className:`gantt-wrap`,ref:Xe,onScroll:()=>Qt(`right`),children:(0,H.jsx)(Ve,{orderedIds:Pt,nodesById:$,displayStatusById:Mt,gitCheckoutById:Rt,registryByNode:e.registry_by_node,gitEnrichment:e.git_enrichment,stackHeaderPx:Ze,depthOffset:It,depths:e.dependency_depths,spans:e.dependency_spans,edges:e.edges,showInheritedEdges:ze,selectedId:c,highlightRowIds:Vt,onSelect:l,onChartBackgroundMouseDown:q?()=>void kt():void 0})})]}):(0,H.jsx)(`p`,{style:{padding:`1rem`},children:`Loading roadmap…`}),(0,H.jsxs)(F.Suspense,{fallback:null,children:[e?u.map((t,n)=>{let r=$[t];if(!r)return null;let i=r.status?.trim()||`Not Started`,a=Mt[t]??i,o=Lt[t]??a,c=M(!!Rt[t],a,o),l=u.length>1;return(0,H.jsx)(ht,{node:r,pmDisplayStatusResolved:Mt[t],allNodes:e.nodes,modalStorageKey:t,stackZIndex:50+n,backdropPassThrough:l,closeOnEscape:n===u.length-1,headerMinTop:I,spawnInitialRect:m[t],editTileMode:v,tileRect:T?.[t]??null,resumeFreeRect:re?.[t]??null,titleBarActive:f===t,onActivate:()=>qt(t),onRectCommit:e=>Gt(t,e),dependencyInheritance:e.dependency_inheritance?.[t],registryByNode:e.registry_by_node,gitEnrichment:e.git_enrichment,prHints:e.pr_hints,onClose:()=>Jt(t),onOpenNode:Kt,onPersisted:()=>void Tt(`Saving task…`,X).catch(e=>s(String(e))),readOnlyCheckout:c},t)}):null,(0,H.jsx)(gt,{open:xe,onClose:()=>{Se(!1),ot()}}),(0,H.jsx)(yt,{open:R,onClose:()=>{z(!1),ot()}}),(0,H.jsx)(vt,{open:Ce,onClose:()=>B(!1)}),(0,H.jsx)(bt,{open:we,onClose:()=>Te(!1)}),(0,H.jsx)(_t,{open:_e,onClose:()=>be(!1),themeMode:We,onThemeModeChange:Ge,highlightDepChain:Be,onHighlightDepChainChange:He,showInheritedDeps:ze,onShowInheritedDepsChange:G,refreshSec:Ie,onRefreshSecChange:Le}),(0,H.jsx)(at,{open:ke,onClose:()=>Me(!1),status:Ne,onRefreshStatus:ut,onPublish:pt,headerMinTop:I})]})]})]})})}(0,I.createRoot)(document.getElementById(`root`)).render((0,H.jsx)(F.StrictMode,{children:(0,H.jsx)(Tt,{})}));
