+++ specy_road/pm_gantt_static/assets/index-Od_6nDuf.js
@@ -1,0 +1,4 @@
+const __vite__mapDeps=(i,m=__vite__mapDeps,d=(m.f||(m.f=["assets/EditModal-BEGw50si.js","assets/rolldown-runtime-Dw2cE7zH.js","assets/dnd-stack-Bjh9ZQ0p.js","assets/markdown-stack-BGy33aUx.js","assets/vendor-Dlt94R7z.js","assets/ModalFrame-CHD6Jw2d.js","assets/MarkdownWorkspace-D3ycvloL.js","assets/pmDisplayStatus-CVVg5boq.js","assets/usePmGuiHandlers-C3CRxUMr.js","assets/pmGuiContext-BJC_xsft.js","assets/ConstitutionDrawer-D9R5EKXk.js","assets/ModalPersistStatusFooter-qwFsXZFm.js","assets/SettingsDrawer-B4M0PjtN.js","assets/toolbarIcons-BtpHV_T4.js","assets/SharedDocsDrawer-D9Z9e_j7.js","assets/VisionDrawer-D0kArwWP.js","assets/WorkNotesDrawer-DwaUPJ_M.js"])))=>i.map(i=>d[i]);
+import{r as e}from"./rolldown-runtime-Dw2cE7zH.js";import{a as t,c as n,d as r,f as i,i as a,l as o,m as s,n as c,o as l,p as u,r as d,s as f,t as p,u as m}from"./dnd-stack-Bjh9ZQ0p.js";import{t as h}from"./react-vendor-CWFJq4fC.js";import{O as g,S as _,_ as v,a as y,c as b,d as x,f as S,g as ee,n as C,o as te,p as ne,r as w,s as T,t as E,v as re,w as ie,y as ae}from"./ModalFrame-CHD6Jw2d.js";import{a as oe,c as D,i as O,n as k,o as A,r as se,s as ce,t as j}from"./pmDisplayStatus-CVVg5boq.js";import{c as M}from"./markdown-stack-BGy33aUx.js";import{a as le,c as ue,d as de,l as fe,n as pe,o as me,s as he,t as ge}from"./toolbarIcons-BtpHV_T4.js";import{n as _e}from"./pmGuiContext-BJC_xsft.js";(function(){let e=document.createElement(`link`).relList;if(e&&e.supports&&e.supports(`modulepreload`))return;for(let e of document.querySelectorAll(`link[rel="modulepreload"]`))n(e);new MutationObserver(e=>{for(let t of e)if(t.type===`childList`)for(let e of t.addedNodes)e.tagName===`LINK`&&e.rel===`modulepreload`&&n(e)}).observe(document,{childList:!0,subtree:!0});function t(e){let t={};return e.integrity&&(t.integrity=e.integrity),e.referrerPolicy&&(t.referrerPolicy=e.referrerPolicy),e.crossOrigin===`use-credentials`?t.credentials=`include`:e.crossOrigin===`anonymous`?t.credentials=`omit`:t.credentials=`same-origin`,t}function n(e){if(e.ep)return;e.ep=!0;let n=t(e);fetch(e.href,n)}})();var N=e(s(),1),P=h(),ve=28,ye=28;function be(e,t){return e?C({left:e.left+ve,top:e.top+ye,width:e.width,height:e.height},{minTop:t}):w({minTop:t})}function xe(e,t,n){let r=new Set(e),i=new Map;n.forEach((e,t)=>i.set(e,t));let a=new Map;for(let n of e){let e=t[n]?.node_key;e&&a.set(e,n)}let o=new Map,s=new Map;for(let t of e)o.set(t,0),s.set(t,[]);for(let n of e){let e=t[n]?.dependencies??[];for(let t of e){let e=a.get(t);!e||!r.has(e)||e===n||(s.get(e).push(n),o.set(n,(o.get(n)??0)+1))}}let c=(e,t)=>(i.get(e)??0)-(i.get(t)??0),l=e.filter(e=>o.get(e)===0).sort(c),u=[];for(;l.length;){let e=l.shift();u.push(e);for(let t of s.get(e)??[])o.set(t,(o.get(t)??0)-1),o.get(t)===0&&(l.push(t),l.sort(c))}if(u.length<e.length){let t=e.filter(e=>!u.includes(e)).sort(c);return[...u,...t]}return u}function Se(e,t){let n=window.innerWidth,r=window.innerHeight,i=t+8,a=e.length;if(a===0)return{};let o=n-16,s=r-i-8,c=(o-6*(a-1))/a,l={};return e.forEach((e,n)=>{l[e]=C({left:8+n*(c+6),top:i,width:c,height:s},{minTop:t})}),l}var Ce=`Complete`,we=`In Progress`;function Te(){try{return!0}catch{return!0}}function F(e){return(e||``).trim().toLowerCase()}function Ee(e){return F(e)===`blocked`}function De(e){return F(e)===`complete`}function Oe(e){return F(e)===`in progress`}function ke(e){let t=new Map;for(let n of e){let e=n.parent_id===void 0||n.parent_id===null||n.parent_id===``?null:String(n.parent_id),r=n.id;t.has(e)||t.set(e,[]),t.get(e).push(r)}return t}function Ae(e,t,n){let r=[],i=new Set;function a(e){if(!i.has(e)){i.add(e);for(let t of n.get(e)??[])a(t);r.push(e)}}for(let n of e){let e=t[n]?.parent_id;(e==null||typeof e==`string`&&e.trim()===``)&&a(n)}for(let t of e)i.has(t)||a(t);return r}function I(e,t,n){let r={};for(let i of e){let e=t[i]??`Not Started`;if(De(e)){r[i]=Ce;continue}if(Ee(e)){r[i]=e;continue}let a=n.get(i);if(!a||a.length===0){r[i]=e;continue}let o=a.some(e=>Oe(r[e]));if(a.every(e=>F(r[e])===`complete`)){r[i]=Ce;continue}if(o){r[i]=we;continue}r[i]=e}return r}function je(e,t,n,r){let i={};for(let r of e)i[r]=j(t[r],n?.[r]);if(!(r?.enabled??Te()))return{...i};let a=ke(e.map(e=>t[e]).filter(e=>!!e));return I(Ae(e,t,a),i,a)}function Me(e,t,n,r){let i=r?.enabled??Te(),a={};for(let r of e)a[r]=j(t[r],n?.[r]);if(!i)return{...a};let o=ke(e.map(e=>t[e]).filter(e=>!!e)),s=I(Ae(e,t,o),a,o),c={...a};for(let n of e){let e=t[n],r=(o.get(n)?.length||0)>0,i=F(a[n]);r&&s[n]===`In Progress`&&i!==`complete`&&i!==`blocked`&&i!==`in progress`&&(c[n]=we),e?.type===`phase`&&s[n]===`Complete`&&(c[n]=Ce)}return c}function Ne(e,t){return!e||e.type!==`phase`||F(e.status||``)===`complete`?!1:F(t)===`complete`}function Pe(e,t){let n=new Set,r=e;for(;r;){for(let e of r.dependencies??[])n.add(e);let e=r.parent_id;if(!e)break;r=t[e]}return n}function Fe(e,t,n){let r=new Set,i=new Set,a=[e];for(;a.length;){let e=a.shift();if(i.has(e))continue;i.add(e);let o=t[e];if(o)for(let e of Pe(o,t)){let i=n[e]??e;!i||!t[i]||(r.add(i),a.push(i))}}return r.delete(e),r}function Ie(e,t){if(e.length===0)return 0;let n=1/0;for(let r of e){let e=t[r]??0;e<n&&(n=e)}return n===1/0?0:n}function Le(e){let t=new Set(e.issues.map(e=>e.code)),{resolved:n,config:r}=e,i=r?.integration_branch??n.integration_branch,a=r?.remote??n.remote;if(t.has(`missing_config_file`))return{tone:`red`,label:`No git-workflow.yaml`,tooltipIntro:`Red means the tracked contract file roadmap/git-workflow.yaml is missing. Add it from the specy-road init project template (version, integration_branch, remote) so CLI defaults and this status match your team.`};if(t.has(`invalid_config`))return{tone:`red`,label:`Invalid git-workflow.yaml`,tooltipIntro:`Red means the file exists but failed schema or YAML validation. Fix the file to match schemas/git-workflow.schema.json, then reload.`};if(t.has(`not_git_repo`))return{tone:`yellow`,label:`Git not connected`,tooltipIntro:`Yellow means this directory is not a git worktree, so branch names and trunk refs cannot be checked. Open the project from a clone or set SPECY_ROAD_REPO_ROOT to the repository root.`};if(t.has(`integration_ref_missing`))return{tone:`yellow`,label:`Fetch ${a}/${i}`,tooltipIntro:`Yellow means the contract is valid and git is connected, but there is no local ref for the integration branch yet (${a}/${i}). Run: git fetch ${a}, or create/checkout that branch locally.`};let o=n.git_branch_current,s=n.git_head_short;return{tone:`green`,label:o!=null&&o.length>0?`${i} · ${a} · ${o}`:`${i} · ${a}${s?` · ${s}`:``}`,tooltipIntro:`Green means roadmap/git-workflow.yaml is valid, this tree is a git checkout, and the integration branch ref exists locally. CLI defaults and PM features use these values.`}}function Re(e,t,n){if(t===void 0||t.enabled!==!0)return e;let{skipped_reason:r,sync_state:i,integration_branch:a,remote:o}=t,s=a,c=o,l=e.tone,u=e.label,d=e.tooltipIntro,f=n&&(r===`dirty_working_tree`||r===`not_on_integration_branch`)?` Merged registry in the PM view can still include claims from ${c}/${s} after git fetch (remote overlay).`:``;return r===`dirty_working_tree`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped (dirty tree)`,d=`${d}\n\nAuto fast-forward of ${s} is skipped because the working tree is not clean. Commit, stash, or discard changes to allow periodic merge --ff-only, or pull ${s} manually.${f}`):r===`not_on_integration_branch`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped (branch)`,d=`${d}\n\nAuto fast-forward runs only when the current branch is the integration branch (${s}).${f}`):r===`not_git_repo`?(l===`green`&&(l=`yellow`),u=`${u} · FF skipped`,d=`${d}\n\nAuto fast-forward requires a git worktree.`):r===`integration_ref_unavailable`?(l===`green`&&(l=`yellow`),u=`${u} · no ${c}/${s} ref`,d=`${d}\n\nRun git fetch ${c} so refs/remotes/${c}/${s} exists.`):i===`diverged`?(l===`green`&&(l=`yellow`),u=`${u} · diverged`,d=`${d}\n\nLocal ${s} and ${c}/${s} have diverged; fast-forward is not possible. Resolve with merge/rebase, then pull.`):i===`ahead_of_remote`&&(l===`green`&&(l=`yellow`),u=`${u} · ahead of ${c}`,d=`${d}\n\nThis branch is ahead of ${c}/${s}; push or reset before others can fast-forward.`),{tone:l,label:u,tooltipIntro:d}}var L=M();function ze({gitWorkflow:e,integrationBranchAutoFf:t,registryRemoteOverlay:n=!1}){if(e===void 0)return(0,L.jsx)(`span`,{className:`app-header-git-workflow app-header-git-workflow--loading`,"aria-hidden":`true`,children:`Git —…`});let r=Re(Le(e),t,n),i=`git-workflow-status-tip`,{resolved:a}=e;return(0,L.jsxs)(`div`,{className:`app-header-git-workflow-tooltip app-header-doc-tooltip`,children:[(0,L.jsx)(`span`,{id:`git-workflow-status-label`,className:`app-header-git-workflow-label app-header-git-workflow-label--${r.tone}`,tabIndex:0,"aria-describedby":i,children:r.label}),(0,L.jsxs)(`div`,{id:i,role:`tooltip`,className:`app-header-doc-tip app-header-doc-tip--wide app-header-doc-tip--git-workflow`,children:[(0,L.jsx)(`p`,{className:`git-workflow-tip-intro`,children:r.tooltipIntro}),(0,L.jsxs)(`div`,{className:`git-workflow-outline-hint`,children:[(0,L.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Outline`}),(0,L.jsxs)(`p`,{className:`git-workflow-outline-hint-text`,children:[`A green left accent on a row means your named git branch matches the`,` `,(0,L.jsx)(`code`,{children:`branch`}),` field for that task in`,` `,(0,L.jsx)(`code`,{children:`roadmap/registry.yaml`}),` in `,(0,L.jsx)(`em`,{children:`this`}),` checkout (not necessarily on other branches).`]})]}),e.issues.length>0?(0,L.jsxs)(`div`,{className:`git-workflow-issues-block`,children:[(0,L.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Details`}),e.issues.map(e=>(0,L.jsxs)(`div`,{className:`git-workflow-issue`,children:[(0,L.jsx)(`strong`,{children:e.message}),(0,L.jsx)(`div`,{className:`git-workflow-issue-detail`,children:e.detail})]},e.code))]}):null,(0,L.jsxs)(`div`,{className:`git-workflow-resolved-block`,children:[(0,L.jsx)(`div`,{className:`git-workflow-tip-section-title`,children:`Resolved`}),(0,L.jsxs)(`ul`,{className:`git-workflow-resolved-list`,children:[(0,L.jsxs)(`li`,{children:[`Integration branch:`,` `,(0,L.jsx)(`code`,{children:a.integration_branch})]}),(0,L.jsxs)(`li`,{children:[`Remote: `,(0,L.jsx)(`code`,{children:a.remote})]}),(0,L.jsxs)(`li`,{children:[`Current branch:`,` `,a.git_branch_current!=null&&a.git_branch_current.length>0?a.git_branch_current:`(detached or unnamed)`]}),a.git_head_short?(0,L.jsxs)(`li`,{children:[`HEAD: `,(0,L.jsx)(`code`,{children:a.git_head_short})]}):null]})]})]})]})}var Be=new Set([`github_pr`,`gitlab_mr`]);function Ve(e){if(!e)return{open:!1,merged:!1,rejected:!1};let t=e.kind||``,n=(e.pr_state||``).toLowerCase(),r=!!e.merged;return n===`merged`||r?{open:!1,merged:!0,rejected:!1}:n===`rejected`||n===`closed`?{open:!1,merged:!1,rejected:!0}:n===`open`||Be.has(t)?{open:!0,merged:!1,rejected:!1}:{open:!1,merged:!1,rejected:!1}}function He(e,t,n,r,i,a){let o=(t||``).trim().toLowerCase(),s=(n||``).trim().toLowerCase(),c=a?.[e],l=Ve(c);return l.merged?`var(--bar-complete)`:l.rejected?`var(--bar-mr-rejected)`:s===`blocked`||o===`blocked`?`var(--bar-blocked)`:l.open?`var(--bar-mr-pending)`:D(e,t,r,i)?`var(--bar-in-progress)`:o===`complete`||s===`complete`?`var(--bar-complete)`:o===`not started`||!t?`var(--bar-not-started)`:A(t)?`var(--bar-in-progress)`:`var(--bar-not-started)`}function Ue(e){let t=e.node?.status,n=He(e.nodeId,e.displayStatus,t,e.registryByNode,e.gitCheckoutById,e.gitEnrichment);if(e.selected)return{fill:`var(--gantt-bar-selected)`,stroke:`rgba(0,0,0,0.15)`,strokeWidth:1};if(e.depHighlight)return{fill:`var(--gantt-dep-chain)`,stroke:`rgba(0,0,0,0.15)`,strokeWidth:1};if(e.node?.type===`gate`)return{fill:`var(--bar-gate)`,stroke:`rgba(0,0,0,0.12)`,strokeWidth:1};let r=n===`var(--bar-in-progress)`&&D(e.nodeId,e.displayStatus,e.registryByNode,e.gitCheckoutById);return{fill:n,stroke:r?`var(--accent)`:`rgba(0,0,0,0.15)`,strokeWidth:r?2:1}}var R=38,z=52,We=.82,B=8;function Ge(e){let t=Math.max(1,e);return t<=1?We*z:(t-1)*z+We*z}function Ke({orderedIds:e,nodesById:t,displayStatusById:n,gitCheckoutById:r,registryByNode:i,gitEnrichment:a,stackHeaderPx:o=52,depthOffset:s=0,depths:c,spans:l={},edges:u,showInheritedEdges:d=!0,selectedId:f,highlightRowIds:p=null,onSelect:m,onChartBackgroundMouseDown:h}){let g=e.length;if(g===0)return null;let _=e=>Math.max(0,(c[e]??0)-s),v=0;for(let t of e){let e=_(t),n=l[t]??1;v=Math.max(v,e+n)}let y=v+2,b=B+y*z+16,x=Math.max(32,o),S=g*R+x,ee=Math.max(16,x-6),C={};e.forEach((e,t)=>{C[e]=t});let te=d?u:u.filter(e=>e.kind!==`inherited`);return(0,L.jsxs)(`svg`,{className:`gantt-svg`,width:b,height:S,role:`img`,"aria-label":`Dependency Gantt`,children:[h?(0,L.jsx)(`rect`,{x:0,y:0,width:b,height:S,fill:`transparent`,style:{pointerEvents:`all`,cursor:`default`},onMouseDown:e=>{e.preventDefault(),h()}}):null,(0,L.jsx)(`text`,{x:B,y:18,className:`axis-title`,pointerEvents:`none`,children:`Dependency step (not calendar time)`}),(0,L.jsx)(`text`,{x:B+280,y:18,className:`gantt-legend`,fontSize:10,fill:`var(--muted)`,pointerEvents:`none`,children:d?`Solid = explicit dep · Dashed = inherited`:`Solid = explicit dependency`}),Array.from({length:y},(e,t)=>(0,L.jsxs)(`g`,{pointerEvents:`none`,children:[(0,L.jsx)(`line`,{x1:B+t*z,y1:x,x2:B+t*z,y2:S,stroke:`var(--border)`,strokeWidth:1}),(0,L.jsx)(`text`,{x:B+t*z+4,y:ee,fontSize:10,fill:`var(--muted)`,children:s+t+1})]},t)),(0,L.jsx)(`line`,{x1:0,y1:x,x2:b,y2:x,stroke:`var(--border)`,strokeWidth:1,pointerEvents:`none`}),e.map((e,t)=>(0,L.jsx)(`line`,{x1:0,y1:x+t*R+R,x2:b,y2:x+t*R+R,stroke:`var(--border)`,strokeOpacity:.5,strokeWidth:1,pointerEvents:`none`},`h-${e}`)),p&&p.size>0?e.map((e,t)=>p.has(e)?(0,L.jsx)(`rect`,{x:0,y:x+t*R,width:b,height:R,fill:`var(--gantt-dep-chain-bg)`,opacity:.35,pointerEvents:`none`},`dep-hi-${e}`):null):null,e.map((e,o)=>{let s=t[e],c=_(e),u=Ge(l[e]??1),d=x+o*R,h=B+c*z,g=f===e,{fill:v,stroke:y,strokeWidth:b}=Ue({nodeId:e,selected:g,depHighlight:!!p?.has(e)&&!g,displayStatus:n?.[e]??s?.status,node:s,registryByNode:i,gitCheckoutById:r,gitEnrichment:a});return(0,L.jsx)(`rect`,{x:h,y:d+6,width:u,height:R-12,rx:3,fill:v,stroke:y,strokeWidth:b,style:{cursor:`pointer`,pointerEvents:`all`},onMouseDown:e=>e.stopPropagation(),onClick:()=>m(e)},`bar-${e}`)}),te.map(({from:e,to:t,kind:n})=>{let r=C[e],i=C[t];if(r===void 0||i===void 0)return null;let a=_(e),o=_(t),s=Ge(l[e]??1),c=B+a*z+s,u=B+o*z,d=x+r*R+R/2,f=x+i*R+R/2,p=(c+u)/2,m=n===`inherited`;return(0,L.jsx)(`path`,{d:`M ${c} ${d} C ${p} ${d}, ${p} ${f}, ${u} ${f}`,fill:`none`,stroke:`var(--accent)`,strokeWidth:m?1:1.25,strokeOpacity:m?.55:.75,strokeDasharray:m?`5 4`:void 0,pointerEvents:`none`},`${e}->${t}-${n??`x`}`)})]})}var qe=e(u(),1),V={pendingIds:new Set,pendingFor:()=>null},Je=(0,N.createContext)(null);function Ye(){return(0,N.useContext)(Je)??V}function H(e){let t=e?.parent_id;return t==null||t===``?null:t}function U(e){return e.map(e=>({...e,dependencies:[...e.dependencies??[]]}))}function Xe(e){let t=new Map;for(let n=0;n<e.length;n++)t.set(e[n].id,n);return t}function W(e,t){let n=[];for(let r=0;r<e.length;r++)H(e[r])===t&&n.push({id:e[r].id,sibling_order:Number(e[r].sibling_order??0),idx:r});return n.sort((e,t)=>e.sibling_order===t.sibling_order?e.id.localeCompare(t.id):e.sibling_order-t.sibling_order),n}function Ze(e,t,n){let r=Xe(e),i=0;for(let t of n){let n=r.get(t);n!=null&&(e[n]={...e[n],sibling_order:i++})}for(let r of W(e,t))n.includes(r.id)||(e[r.idx]={...e[r.idx],sibling_order:i++})}function Qe(e,t){return Ze(e,t.parentId,t.orderedChildIds),e}function $e(e,t){let n=e.findIndex(e=>e.node_key===t.nodeKey);if(n<0)return e;let r=e[n];e[n]={...r,parent_id:t.newParentId};let i=W(e,t.newParentId).filter(e=>e.id!==r.id).map(e=>e.id),a=Math.max(0,Math.min(t.newIndex,i.length)),o=[...i.slice(0,a),r.id,...i.slice(a)];return Ze(e,t.newParentId,o),e}function et(e,t){let n=e.findIndex(e=>e.id===t.nodeId);if(n<0)return e;let r=e[n],i=W(e,H(r)).map(e=>e.id),a=i.indexOf(t.nodeId);if(a<=0)return e;let o=i[a-1];return $e(e,{kind:`move`,nodeKey:r.node_key,newParentId:o,newIndex:W(e,o).length})}function tt(e,t){let n=e.findIndex(e=>e.id===t.nodeId);if(n<0)return e;let r=e[n],i=H(r);if(i==null)return e;let a=e.findIndex(e=>e.id===i);if(a<0)return e;let o=H(e[a]),s=W(e,o).map(e=>e.id),c=s.indexOf(i);return $e(e,{kind:`move`,nodeKey:r.node_key,newParentId:o,newIndex:c<0?s.length:c+1})}function nt(e,t){let n=e.findIndex(e=>e.id===t.nodeId);return n<0||(e[n]={...e[n],dependencies:[...t.explicitNodeKeys]}),e}function rt(e,t){let n=e.findIndex(e=>e.id===t.referenceNodeId);if(n<0)return e;let r=e[n],i={...t.placeholder,parent_id:H(r)},a=[...e,i],o=W(a,H(r)).map(e=>e.id).filter(e=>e!==i.id),s=o.indexOf(t.referenceNodeId),c=t.position===`above`?s:s+1,l=[...o.slice(0,c),i.id,...o.slice(c)];return Ze(a,H(r),l),a}function it(e,t){let n=new Set,r=t=>{n.add(t);for(let n of e)n.parent_id===t&&r(n.id)};return r(t.nodeId),e.filter(e=>!n.has(e.id))}function at(e){let t=new Map;for(let n of e){let e=H(n),r=t.get(e)??[];r.push(n),t.set(e,r)}for(let e of t.values())e.sort((e,t)=>{let n=Number(e.sibling_order??0),r=Number(t.sibling_order??0);return n===r?e.id.localeCompare(t.id):n-r});let n=[],r=(e,i)=>{for(let a of t.get(e)??[])n.push({node:a,depth:i}),r(a.id,i+1)};return r(null,0),n}function ot(e,t){let n=at(t);return{...e,nodes:t,tree:n.map((e,t)=>({id:e.node.id,outline_depth:e.depth,row_index:t})),ordered_ids:n.map(e=>e.node.id),row_depths:n.map(e=>e.depth)}}function st(e,t){if(t.length===0)return e;let n=U(e.nodes);for(let e of t)switch(e.kind){case`reorder`:n=Qe(n,e);break;case`move`:n=$e(n,e);break;case`indent`:n=et(n,e);break;case`outdent`:n=tt(n,e);break;case`dep`:n=nt(n,e);break;case`add`:n=rt(n,e);break;case`delete`:n=it(n,e);break}return ot(e,n)}function ct(e){return{id:`__pending__:${e.token}`,node_key:`pending-${e.token}`,type:e.type,title:e.title,status:`Not Started`,parent_id:e.parentId,dependencies:[],sibling_order:0}}function lt(e){return e.startsWith(`__pending__:`)}function ut(e){let t=e?.parent_id;return t==null||t===``?null:t}function dt(e,t,n){if(t===n)return!0;let r=new Set,i=ut(e[t]);for(;i;){if(i===n)return!0;if(r.has(i))return!1;r.add(i),i=ut(e[i])}return!1}function G(e,t,n){return e.includes(n)?e.filter(e=>dt(t,e,n)):[]}var ft=`Title and planning are read-only while this task is in active development (in progress, open or merged merge request, or this checkout matches the registered branch).`,K=e=>{let t=n(e);return t.length>0?t:f(e)},pt=5;function mt(e,t,n,r){let i=Ne(e,n),a=r!==t,o=i&&r.trim().toLowerCase()===`complete`&&t.trim().toLowerCase()!==`complete`,s=[];return i&&s.push(`Display: all descendants complete (roadmap status unchanged).`),a&&!o&&s.push(`PM view reflects active registration or Git remote state; the roadmap file may still list another status until it is updated.`),s.length>0?s.join(`
+
+`):void 0}function q(e){let t=e?.parent_id;return t==null||t===``?null:t}function ht(e,t,n){let r=e[n]?.parent_id;for(;r;){if(r===t)return!0;r=e[r]?.parent_id??null}return!1}function gt(e,t,n){return t?t===e?!0:ht(n,e,t):!1}function _t(e,t,n,r,i){let a=r.filter(t=>q(i[t])===e);if(!a.includes(t)||!a.includes(n))return null;let o=a.filter(e=>e!==t),s=o.indexOf(n);return s<0?null:[...o.slice(0,s),t,...o.slice(s)]}function vt(e,t,n,r,i){let a=r.filter(t=>q(i[t])===e);if(!a.includes(t)||!a.includes(n))return null;let o=a.filter(e=>e!==t),s=o.indexOf(n);if(s<0)return null;let c=s+1;return[...o.slice(0,c),t,...o.slice(c)]}function yt({nodeId:e}){let{setNodeRef:t,isOver:n}=o({id:`into:${e}`});return(0,L.jsx)(`span`,{ref:t,className:n?`into-drop into-drop-active`:`into-drop`,title:`Drop here to make this row the new parent (last child)`,children:`⎘`})}function J(){let{setNodeRef:e,isOver:t}=o({id:`into:__root__`});return(0,L.jsx)(`th`,{ref:e,colSpan:pt,className:t?`into-root into-root-active`:`into-root`,title:`Drop here to move to top level (append as last root)`,children:`Top level`})}function Y({targetId:e,onInsertClick:t,depEditActive:n,hidden:r}){let{setNodeRef:i,isOver:a}=o({id:`before:${e}`});return(0,L.jsx)(`tr`,{ref:i,className:`outline-gap-tr`,style:r?{opacity:0,pointerEvents:`none`}:void 0,children:(0,L.jsx)(`td`,{className:a?`outline-gap-cell outline-gap-active`:`outline-gap-cell`,colSpan:pt,title:`Drop here to insert before this row (same parent as this row)`,children:(0,L.jsx)(`button`,{type:`button`,className:`outline-gap-insert-btn`,"aria-label":`Add task above`,disabled:n,onPointerDown:e=>e.stopPropagation(),onClick:n=>{n.stopPropagation(),t(e)},children:`+`})})})}function bt(e){let{variant:t,id:n,node:r,outlineDepth:i,meta:a,statusText:o,statusCellTitle:s,devText:c,depCellText:l,depEditId:u,titleEditing:d,titleDraft:f,onTitleDraftChange:p,onTitleBlur:m,onTitleKeyDown:h,titleInputRef:g,onDepCellClick:_,depCellRef:v,devCellTitle:y,titleEditLocked:b,rowClass:x,gitCheckoutTitle:S,trRef:ee,trStyle:C,dndListeners:te,dndAttributes:ne,onRowDoubleClick:w,onTitleClick:T,onTitlePointerDown:E,onTitlePointerMove:re,onTitlePointerEnd:ie,onIdClick:ae,onStatusDevClick:oe}=e,D=u===n,O=t===`dragPreview`;return(0,L.jsxs)(`tr`,{ref:ee,style:C,className:x||void 0,title:S,...O?{}:te??{},...O?{}:ne??{},onDoubleClick:w,children:[(0,L.jsxs)(`td`,{className:`outline-id`,onClick:O?void 0:ae,children:[(0,L.jsx)(`span`,{className:`outline-id-text`,children:lt(r.id)?`…`:r.id}),O?null:(0,L.jsx)(yt,{nodeId:r.id})]}),(0,L.jsxs)(`td`,{className:`outline-title`,title:b?ft:void 0,onClick:O?void 0:T,onPointerDown:O?void 0:E,onPointerMove:O?void 0:re,onPointerUp:O?void 0:ie,onPointerCancel:O?void 0:ie,onPointerLeave:O?void 0:ie,style:{paddingLeft:`${i*12}px`},children:[!O&&d?(0,L.jsx)(`input`,{ref:g,className:`outline-title-input`,value:f,onChange:e=>p(e.target.value),onBlur:m,onKeyDown:h,onClick:e=>e.stopPropagation(),onPointerDownCapture:e=>e.stopPropagation(),onDoubleClick:e=>e.stopPropagation()}):(0,L.jsxs)(`div`,{className:`outline-title-row`,children:[(0,L.jsx)(`div`,{className:`outline-title-text-wrap`,children:(0,L.jsx)(`div`,{children:r.title})}),b?(0,L.jsx)(`span`,{className:`outline-pm-lock`,title:ft,"aria-label":`Read-only`,children:(0,L.jsxs)(`svg`,{className:`outline-pm-lock-icon`,width:`14`,height:`14`,viewBox:`0 0 24 24`,fill:`none`,xmlns:`http://www.w3.org/2000/svg`,"aria-hidden":`true`,focusable:`false`,children:[(0,L.jsx)(`rect`,{x:`3`,y:`11`,width:`18`,height:`11`,rx:`2`,stroke:`currentColor`,strokeWidth:`2`}),(0,L.jsx)(`path`,{d:`M7 11V7a5 5 0 0 1 10 0v4`,stroke:`currentColor`,strokeWidth:`2`,strokeLinecap:`round`})]})}):null]}),a?(0,L.jsx)(`div`,{className:`outline-meta`,children:a}):null]}),(0,L.jsx)(`td`,{className:`outline-col-status`,title:s,onClick:O?void 0:oe,children:o}),(0,L.jsx)(`td`,{className:`outline-col-dev`,title:y,onClick:O?void 0:oe,children:c}),(0,L.jsx)(`td`,{ref:v,className:D?`outline-col-dep outline-col-dep-active`:`outline-col-dep`,title:`Click to choose which features this item depends on (explicit)`,onClick:O?void 0:e=>{e.preventDefault(),e.stopPropagation(),_()},children:(0,L.jsx)(`span`,{className:`outline-dep-cell-text`,children:l})})]})}function xt({id:e,node:t,outlineDepth:n,selected:r,meta:a,statusText:o,statusCellTitle:s,devText:l,depCellText:u,depEditId:d,isDepCandidate:f,titleEditing:p,titleDraft:m,onTitleDraftChange:h,onTitleBlur:g,onTitleKeyDown:_,titleInputRef:v,onBeginTitleEdit:y,onSelectRow:b,onOpenModal:x,onDepRowBodyClick:S,onDepCellClick:ee,dragDisabled:C,depCellRef:te,isGitCheckoutRow:ne,devCellTitle:w,titleEditLocked:T,rowSourceHidden:E}){let{attributes:re,listeners:ie,setNodeRef:ae,transform:oe,transition:D}=c({id:e,disabled:C}),O={transform:i.Transform.toString(oe),transition:D,opacity:+!E,pointerEvents:E?`none`:void 0},k=(0,N.useRef)(null),A=(0,N.useRef)(null),se=(0,N.useRef)(null),ce=(0,N.useRef)(!1),j=()=>{A.current&&=(clearTimeout(A.current),null),se.current=null},M=e=>{d||p||T||e.button===0&&(j(),se.current={x:e.clientX,y:e.clientY},A.current=setTimeout(()=>{A.current=null,se.current=null,ce.current=!0,k.current&&=(clearTimeout(k.current),null),r||b(),y()},500))},le=e=>{if(!A.current||!se.current)return;let t=e.clientX-se.current.x,n=e.clientY-se.current.y;t*t+n*n>100&&j()},ue=()=>{j()},de=()=>{if(ce.current){ce.current=!1;return}if(d){S();return}k.current&&clearTimeout(k.current),k.current=setTimeout(()=>{k.current=null,b()},220)},fe=e=>{e.preventDefault(),!d&&(p||(j(),k.current&&=(clearTimeout(k.current),null),x()))},pe=()=>{k.current&&clearTimeout(k.current),k.current=null,b()},me=()=>{d?S():b()},he=Ye().pendingFor(e),ge=he?he.phase===`fail`?`outline-row-pending outline-row-pending--fail`:he.phase===`settling`?`outline-row-pending outline-row-pending--settling`:`outline-row-pending`:``;return(0,L.jsx)(bt,{id:e,node:t,outlineDepth:n,selected:r,meta:a,statusText:o,statusCellTitle:s,devText:l,depCellText:u,depEditId:d,isDepCandidate:f,titleEditing:p,titleDraft:m,onTitleDraftChange:h,onTitleBlur:g,onTitleKeyDown:_,titleInputRef:v,onBeginTitleEdit:y,onSelectRow:b,onOpenModal:x,onDepRowBodyClick:S,onDepCellClick:ee,dragDisabled:C,depCellRef:te,isGitCheckoutRow:ne,devCellTitle:w,titleEditLocked:T,variant:`interactive`,rowClass:[r?`selected`:``,d===e?`dep-edit-row`:``,f?`dep-candidate-row`:``,ne?`outline-row-git-current`:``,t.type===`gate`?`outline-row-gate`:``,ge].filter(Boolean).join(` `),gitCheckoutTitle:ne?`Named branch matches the branch field in roadmap/registry.yaml for this row in this checkout.`:void 0,trRef:ae,trStyle:O,dndListeners:ie,dndAttributes:re,onRowDoubleClick:fe,onTitleClick:de,onTitlePointerDown:M,onTitlePointerMove:le,onTitlePointerEnd:ue,onIdClick:pe,onStatusDevClick:me})}function St({orderedIds:e,nodesById:n,rowDepths:i,selectedId:o,prHints:s,gitEnrichment:c,dependencyInheritance:u,registryByNode:f,gitBranchCurrent:h,gitUserName:g,depEditId:_,depDraftKeys:y,onToggleDepCandidate:b,onDepCellActivate:x,onApplyDepEdit:S,onCancelDepEdit:ee,onClearDepDraft:C,onSelect:te,onDoubleClick:ne,performRoadmapMutation:w,performOptimisticMutation:T,onMutationError:E,onGapInsert:re,displayStatusById:D,outlineStatusById:k,reorderLocked:A=!1,interactionLocked:j=!1}){let[M,le]=(0,N.useState)(null),[ue,de]=(0,N.useState)(``),fe=(0,N.useRef)(null),pe=r(m(l,{activationConstraint:{distance:6}})),me=(0,N.useMemo)(()=>{let e=new Map;for(let t of Object.values(n))t?.node_key&&e.set(t.node_key,t.id);return e},[n]),he=e=>{let t=c[e];if(t?.kind===`registry`||t?.kind===`remote_tip`)return``;if(t?.kind===`github_pr`||t?.kind===`gitlab_mr`){let e=t.assignees;return[e?.length?`A: ${e.join(`, `)}`:``].filter(Boolean).join(` · `)||t.hint_line||``}return t?.hint_line?String(t.hint_line):s[e]?s[e].replace(/<br>/g,` · `):``},ge=e=>{let t=me.get(e);if(t)return t;for(let t of Object.values(n))if(t?.node_key===e)return t.id},_e=e=>{if(!n[e])return`—`;if(_===e){let e=[...y].map(e=>ge(e)).filter(Boolean);return e.length?[...new Set(e)].sort().join(`, `):`—`}let t=u?.[e]?.explicit??[];return t.length?t.join(`, `):`—`},P=(0,N.useCallback)(async()=>{if(!M)return;if(ce(M,f,h)){le(null);return}let e=n[M];if(!e){le(null);return}let t=ue.trim();if(t&&t!==e.title)try{await w(`Saving title…`,()=>ae(M,[{key:`title`,value:t}]))}catch(e){console.error(e),E(String(e))}le(null)},[M,ue,n,w,E,f,h]),ve=(0,N.useCallback)(()=>{le(null)},[]);(0,N.useEffect)(()=>{if(!M)return;let e=fe.current;e&&(e.focus(),e.select())},[M]),(0,N.useEffect)(()=>{if(!M)return;let e=M,t=window.setInterval(()=>{if(ce(e,f,h)){le(null);return}let t=n[e];if(!t)return;let r=ue.trim();r&&r!==t.title&&w(`Saving title…`,()=>ae(e,[{key:`title`,value:r}])).catch(e=>E(String(e)))},2500);return()=>window.clearInterval(t)},[M,ue,n,w,E,f,h]);let ye=(0,N.useCallback)(e=>{if(M){(async()=>{await P(),x(e)})();return}x(e)},[M,P,x]),be=e=>{if(e.key===`Escape`){e.preventDefault(),ve();return}e.key===`Enter`&&(e.preventDefault(),P())},xe=e=>{if(_||j||ce(e,f,h))return;let t=n[e];t&&(le(e),de(t.title))},Se=(0,N.useCallback)(()=>w(`Refreshing roadmap…`,async()=>{}),[w]),Ce=async(t,r)=>{if(t===r)return;let i=n[t],a=n[r];if(!i||!a)return;let o=q(a);if(o!==null&&gt(t,o,n))return;let s=e.filter(e=>q(n[e])===o&&e!==t).indexOf(r);if(!(s<0)){if(q(i)===o){let i=_t(o,t,r,e,n);if(!i?.length)return;try{await T(`Updating outline…`,{kind:`reorder`,parentId:o,orderedChildIds:i},i,`reorder`,()=>ie(o,i))}catch(e){console.error(e),E(String(e)),await Se()}return}try{await T(`Updating outline…`,{kind:`move`,nodeKey:i.node_key,newParentId:o,newIndex:s},[t],`move`,()=>v(i.node_key,o,s))}catch(e){console.error(e),E(String(e)),await Se()}}},we=async(t,r)=>{if(t===r)return;let i=n[t],a=n[r];if(!i||!a)return;let o=q(a);if(o!==null&&gt(t,o,n))return;let s=e.filter(e=>q(n[e])===o&&e!==t).indexOf(r);if(s<0)return;let c=s+1;if(q(i)===o){let i=vt(o,t,r,e,n);if(!i?.length)return;try{await T(`Updating outline…`,{kind:`reorder`,parentId:o,orderedChildIds:i},i,`reorder`,()=>ie(o,i))}catch(e){console.error(e),E(String(e)),await Se()}return}try{await T(`Updating outline…`,{kind:`move`,nodeKey:i.node_key,newParentId:o,newIndex:c},[t],`move`,()=>v(i.node_key,o,c))}catch(e){console.error(e),E(String(e)),await Se()}},Te=(0,N.useRef)(0);(0,N.useEffect)(()=>{let e=e=>{Te.current=e.clientY};return window.addEventListener(`pointermove`,e,{passive:!0}),()=>window.removeEventListener(`pointermove`,e)},[]);let F=async t=>{if(Oe(null),_)return;let{active:r,over:i}=t;if(!i)return;let a=String(r.id),o=String(i.id),s=n[a];if(!s)return;if(o.startsWith(`into:`)){let t=o.slice(5),r=t===`__root__`?null:t;if(r===a)return;if(r&&ht(n,a,r)){await Se();return}if(gt(a,r,n))return;let i=e.filter(e=>q(n[e])===r&&e!==a).length;try{await T(`Updating outline…`,{kind:`move`,nodeKey:s.node_key,newParentId:r,newIndex:i},[a],`move`,()=>v(s.node_key,r,i))}catch(e){console.error(e),E(String(e)),await Se()}return}let c;if(c=o.startsWith(`before:`)?o.slice(7):o,a===c||!n[c])return;let l=!1;if(!o.startsWith(`before:`)&&i?.rect){let e=i.rect,t=e.top+e.height/2;l=Te.current>t}l?await we(a,c):await Ce(a,c)},Ee=!!_||A||j,[De,Oe]=(0,N.useState)(null),ke=(0,N.useMemo)(()=>De?G(e,n,De):[],[De,e,n]),Ae=ke[0]??null,I=(0,N.useMemo)(()=>new Set(ke),[ke]),je=e=>{Oe(String(e.active.id))},Me=()=>{Oe(null)},Ne=(0,N.useRef)(null),Pe=(0,N.useRef)(null),[Fe,Ie]=(0,N.useState)(null);(0,N.useLayoutEffect)(()=>{if(!_){Ie(null);return}let e=Ne.current;if(!e){Ie(null);return}let t=()=>{let t=e.getBoundingClientRect(),n=Pe.current?.offsetWidth??0,r=t.right+8;n>0&&r+n>window.innerWidth-8&&(r=t.left-8-n),r<8&&(r=8),n>0&&r+n>window.innerWidth-8&&(r=Math.max(8,window.innerWidth-8-n)),Ie({position:`fixed`,top:t.top+t.height/2,left:r,transform:`translateY(-50%)`,zIndex:42,visibility:`visible`,pointerEvents:`auto`})};t();let n=requestAnimationFrame(()=>t()),r=e.closest(`.outline-wrap`);window.addEventListener(`resize`,t),r?.addEventListener(`scroll`,t,{passive:!0});let i=new ResizeObserver(t);return i.observe(e),r&&i.observe(r),()=>{cancelAnimationFrame(n),window.removeEventListener(`resize`,t),r?.removeEventListener(`scroll`,t),i.disconnect()}},[_,e,y]);let Le=(e,t)=>{let r=n[e];if(!r)return null;let a=r?.node_key,l=!!_&&!!a&&y.has(a)&&_!==e,u=ce(e,f,h),d=(r?.status)?.trim()||`Not Started`,p=D==null?d:D[e]??d,m=k==null?p:k[e]??p,v=k==null?D==null?r?.status||`—`:p:m,x=mt(r,d,p,m),S=se(u,p,m);return{id:e,node:r,outlineDepth:i[t]??0,selected:o===e,meta:he(e),statusText:v,statusCellTitle:x,devText:oe(e,f,c,h,g),depCellText:_e(e),depEditId:_,isGitCheckoutRow:u,devCellTitle:O(e,f,c,s),titleEditLocked:S,isDepCandidate:l,titleEditing:M===e,titleDraft:ue,onTitleDraftChange:de,onTitleBlur:()=>void P(),onTitleKeyDown:be,titleInputRef:fe,onBeginTitleEdit:()=>xe(e),dragDisabled:Ee||lt(e),onSelectRow:()=>te(e),onOpenModal:()=>ne(e),onDepRowBodyClick:()=>{if(_){if(e===_){te(e);return}b(e)}},onDepCellClick:()=>ye(e),depCellRef:_===e?Ne:void 0}};return(0,L.jsxs)(L.Fragment,{children:[_==null?null:(0,qe.createPortal)((0,L.jsxs)(`div`,{ref:Pe,className:`dep-edit-toolbar-portal`,role:`toolbar`,"aria-label":`Dependency edit actions`,style:Fe??{position:`fixed`,left:-9999,top:0,visibility:`hidden`,pointerEvents:`none`,zIndex:42},children:[(0,L.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn dep-edit-floating-save`,onClick:e=>{e.preventDefault(),e.stopPropagation(),S()},children:`Save`}),(0,L.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn`,onClick:e=>{e.preventDefault(),e.stopPropagation(),ee()},children:`Cancel`}),(0,L.jsx)(`button`,{type:`button`,className:`dep-edit-floating-btn`,disabled:y.size===0,title:`Remove all prerequisites from the draft`,onClick:e=>{e.preventDefault(),e.stopPropagation(),C()},children:`Clear`})]}),document.body),(0,L.jsx)(a,{sensors:pe,collisionDetection:K,onDragStart:je,onDragEnd:F,onDragCancel:Me,children:(0,L.jsxs)(L.Fragment,{children:[(0,L.jsxs)(`table`,{className:`outline-table${_?` dep-edit-mode`:``}`,children:[(0,L.jsxs)(`thead`,{children:[(0,L.jsx)(`tr`,{children:(0,L.jsx)(J,{})}),(0,L.jsxs)(`tr`,{children:[(0,L.jsx)(`th`,{className:`outline-id`,children:`ID`}),(0,L.jsx)(`th`,{className:`outline-title`,children:`Task`}),(0,L.jsx)(`th`,{className:`outline-col-status`,children:`Status`}),(0,L.jsx)(`th`,{className:`outline-col-dev`,children:`Dev`}),(0,L.jsx)(`th`,{className:`outline-col-dep`,children:`Dependency`})]})]}),(0,L.jsx)(`tbody`,{children:(0,L.jsx)(p,{items:e,strategy:d,children:e.map((e,t)=>{let n=Le(e,t);return n?(0,L.jsxs)(N.Fragment,{children:[(0,L.jsx)(Y,{targetId:e,depEditActive:!!_,onInsertClick:re,hidden:!!(De&&Ae&&I.has(e)&&e!==Ae)}),(0,L.jsx)(xt,{...n,rowSourceHidden:!!(De&&I.has(e))})]},e):null})})})]}),(0,L.jsx)(t,{dropAnimation:null,children:De&&ke.length>0?(0,L.jsx)(`table`,{className:`outline-table outline-drag-overlay-table`,children:(0,L.jsx)(`tbody`,{children:ke.map(t=>{let n=e.indexOf(t);if(n<0)return null;let r=Le(t,n);if(!r)return null;let i=[r.selected?`selected`:``,r.depEditId===t?`dep-edit-row`:``,r.isDepCandidate?`dep-candidate-row`:``,r.isGitCheckoutRow?`outline-row-git-current`:``,r.node.type===`gate`?`outline-row-gate`:``].filter(Boolean).join(` `);return(0,L.jsx)(bt,{...r,variant:`dragPreview`,rowClass:i,gitCheckoutTitle:r.isGitCheckoutRow?`Named branch matches the branch field in roadmap/registry.yaml for this row in this checkout.`:void 0,trStyle:{cursor:`grabbing`}},t)})})}):null})]})})]})}var Ct=`roadmap: update plan`;function wt({open:e,onClose:t,status:n,onRefreshStatus:r,onPublish:i,headerMinTop:a,zIndex:o=200}){let[s,c]=(0,N.useState)(Ct),[l,u]=(0,N.useState)(!1),[d,f]=(0,N.useState)(null);if((0,N.useEffect)(()=>{e&&(f(null),c(Ct),r())},[e,r]),!e)return null;let p=!!n?.can_publish&&!l,m=!!n?.blocked,h=n?.detail??null,g=async()=>{let e=s.trim();if(!e){f(`Enter a short description for this update.`);return}f(null),u(!0);try{await i(e),t()}catch(e){f(String(e))}finally{u(!1)}};return(0,L.jsxs)(E,{title:`Publish roadmap changes`,titleId:`publish-roadmap-title`,onClose:t,storageKey:`publish-roadmap`,minTop:a,zIndex:o,bodyClassName:`modal-body--publish`,footer:(0,L.jsxs)(`div`,{className:`modal-footer-actions`,children:[(0,L.jsx)(`button`,{type:`button`,className:`publish-modal-btn publish-modal-btn--ghost`,disabled:l,onClick:t,children:`Cancel`}),(0,L.jsx)(`button`,{type:`button`,className:`publish-modal-btn publish-modal-btn--primary`,disabled:!p,onClick:()=>void g(),children:l?`Publishing…`:`Publish`})]}),children:[(0,L.jsx)(`p`,{className:`outline-meta`,children:`Saves your roadmap, planning, and governance files to git and pushes to the remote so the team can see them. Your edits are already saved on this computer; publishing shares them.`}),n?(0,L.jsxs)(`div`,{className:`publish-roadmap-meta`,children:[(0,L.jsxs)(`div`,{children:[(0,L.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Branch`}),` `,(0,L.jsx)(`code`,{children:n.current_branch??`—`})]}),n.upstream?(0,L.jsxs)(`div`,{children:[(0,L.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Upstream`}),` `,(0,L.jsx)(`code`,{children:n.upstream})]}):null]}):null,m&&h?(0,L.jsxs)(`div`,{className:`publish-roadmap-blocked`,role:`alert`,children:[(0,L.jsx)(`p`,{children:h}),n&&n.out_of_scope_paths.length>0?(0,L.jsx)(`ul`,{className:`publish-roadmap-path-list`,children:n.out_of_scope_paths.slice(0,12).map(e=>(0,L.jsx)(`li`,{children:(0,L.jsx)(`code`,{children:e})},e))}):null,(0,L.jsx)(`p`,{className:`outline-meta`,children:`Ask a developer if you are unsure how to fix this.`})]}):null,!m&&n?.scope_dirty&&n.scope_paths.length>0?(0,L.jsxs)(`div`,{className:`publish-roadmap-scope`,children:[(0,L.jsx)(`span`,{className:`publish-roadmap-meta-label`,children:`Changed files`}),(0,L.jsx)(`ul`,{className:`publish-roadmap-path-list`,children:n.scope_paths.slice(0,20).map(e=>(0,L.jsx)(`li`,{children:(0,L.jsx)(`code`,{children:e})},e))})]}):null,(0,L.jsx)(`label`,{className:`publish-roadmap-label`,htmlFor:`publish-message`,children:`Summary for teammates`}),(0,L.jsx)(`input`,{id:`publish-message`,type:`text`,className:`publish-roadmap-input`,value:s,onChange:e=>c(e.target.value),disabled:l,autoComplete:`off`,spellCheck:!0,maxLength:500}),d?(0,L.jsx)(`p`,{className:`modal-review-error`,role:`alert`,children:d}):null]})}var X={splitPct:`pmGanttSplitPct`,refreshSec:`pmGanttRefreshSec`,showInheritedDeps:`pmGanttShowInheritedDeps`,highlightDepChain:`pmGanttHighlightDepChain`,themeMode:`pmGanttThemeMode`};function Tt(e,t){return t?`${e}:${t}`:e}function Et(e,t){try{let n=Tt(e,t),r=localStorage.getItem(n);return r===null&&t&&(r=localStorage.getItem(e),r!==null&&localStorage.setItem(n,r)),r}catch{return null}}function Dt(e,t,n){try{localStorage.setItem(Tt(e,t),n)}catch{}}function Ot(e){try{return localStorage.getItem(e)}catch{return null}}function kt(){let e=Promise.resolve(void 0);return{enqueue(t){let n=e.then(()=>t());return e=n.catch(()=>{}),n}}}function At(){let e=(0,N.useRef)(null);e.current||=kt();let[t,n]=(0,N.useState)(0),[r,i]=(0,N.useState)(null),a=(0,N.useCallback)((t,r)=>e.current.enqueue(async()=>{i(t),n(e=>e+1);try{await r()}finally{n(e=>{let t=e-1;return t===0&&i(null),t})}}),[]);return{busy:t>0,busyLabel:r,runRoadmapAction:a}}async function jt(e,t){let n=!1;for(;;)try{return await e(),`ok`}catch(e){if(!(e instanceof y))throw e;if(e.retryable&&!n){n=!0,await t();continue}return await t(),`conflict`}}function Mt({api:e,children:t}){return(0,L.jsx)(Je.Provider,{value:e,children:t})}var Nt={fail:3,active:2,settling:1};function Pt(e){let t=new Map,n=new Map,r=e.schedule??((e,t)=>setTimeout(e,t)),i=e.cancel??(e=>clearTimeout(e)),a=e=>{let t=n.get(e);t!=null&&(i(t),n.delete(e))},o=n=>{a(n),t.delete(n)&&e.onChange()},s=(i,s,c)=>{let l=t.get(i);l&&(t.set(i,{...l,phase:s}),a(i),n.set(i,r(()=>o(i),c)),e.onChange())},c=()=>{let e=new Set,n=new Map;for(let r of t.values())for(let t of r.ids){e.add(t);let i=n.get(t);(!i||Nt[r.phase]>Nt[i.phase])&&n.set(t,{kind:r.kind,phase:r.phase})}return{ids:e,byId:n}};return{get pendingIds(){return c().ids},pendingFor(e){return c().byId.get(e)??null},begin(n,r,i){a(n),t.set(n,{ids:[...r],kind:i,phase:`active`,startedAt:Date.now()}),e.onChange()},end(e){s(e,`settling`,250)},fail(e){s(e,`fail`,400)},dispose(){for(let e of n.values())i(e);n.clear(),t.clear()}}}function Ft(){return typeof crypto<`u`&&typeof crypto.randomUUID==`function`?crypto.randomUUID():`pm-${Math.random().toString(36).slice(2)}-${Date.now().toString(36)}`}function It(){let[e,t]=(0,N.useState)(0),n=(0,N.useRef)(null);n.current||=Pt({onChange:()=>t(e=>e+1|0)});let r=n.current;(0,N.useEffect)(()=>()=>{r.dispose()},[r]);let i=(0,N.useMemo)(()=>({pendingIds:r.pendingIds,pendingFor:e=>r.pendingFor(e)}),[r,e]),a=(0,N.useCallback)((e,t,n)=>r.begin(e,t,n),[r]),o=(0,N.useCallback)(e=>r.end(e),[r]),s=(0,N.useCallback)(e=>r.fail(e),[r]);return(0,N.useMemo)(()=>({...i,begin:a,end:o,fail:s}),[i,a,o,s])}var Z=`modulepreload`,Lt=function(e){return`/`+e},Q={},$=function(e,t,n){let r=Promise.resolve();if(t&&t.length>0){let e=document.getElementsByTagName(`link`),i=document.querySelector(`meta[property=csp-nonce]`),a=i?.nonce||i?.getAttribute(`nonce`);function o(e){return Promise.all(e.map(e=>Promise.resolve(e).then(e=>({status:`fulfilled`,value:e}),e=>({status:`rejected`,reason:e}))))}r=o(t.map(t=>{if(t=Lt(t,n),t in Q)return;Q[t]=!0;let r=t.endsWith(`.css`),i=r?`[rel="stylesheet"]`:``;if(n)for(let n=e.length-1;n>=0;n--){let i=e[n];if(i.href===t&&(!r||i.rel===`stylesheet`))return}else if(document.querySelector(`link[href="${t}"]${i}`))return;let o=document.createElement(`link`);if(o.rel=r?`stylesheet`:Z,r||(o.as=`script`),o.crossOrigin=``,o.href=t,a&&o.setAttribute(`nonce`,a),document.head.appendChild(o),r)return new Promise((e,n)=>{o.addEventListener(`load`,e),o.addEventListener(`error`,()=>n(Error(`Unable to preload CSS for ${t}`)))})}))}function i(e){let t=new Event(`vite:preloadError`,{cancelable:!0});if(t.payload=e,window.dispatchEvent(t),!t.defaultPrevented)throw e}return r.then(t=>{for(let e of t||[])e.status===`rejected`&&i(e.reason);return e().catch(i)})},Rt=(0,N.lazy)(()=>$(()=>import(`./EditModal-BEGw50si.js`).then(e=>({default:e.EditModal})),__vite__mapDeps([0,1,2,3,4,5,6,7,8,9]))),zt=(0,N.lazy)(()=>$(()=>import(`./ConstitutionDrawer-D9R5EKXk.js`).then(e=>({default:e.ConstitutionDrawer})),__vite__mapDeps([10,1,2,3,5,6,11,8,9]))),Bt=(0,N.lazy)(()=>$(()=>import(`./SettingsDrawer-B4M0PjtN.js`).then(e=>({default:e.SettingsDrawer})),__vite__mapDeps([12,1,2,3,5,13]))),Vt=(0,N.lazy)(()=>$(()=>import(`./SharedDocsDrawer-D9Z9e_j7.js`).then(e=>({default:e.SharedDocsDrawer})),__vite__mapDeps([14,1,2,3,5,8,9]))),Ht=(0,N.lazy)(()=>$(()=>import(`./VisionDrawer-D0kArwWP.js`).then(e=>({default:e.VisionDrawer})),__vite__mapDeps([15,1,2,3,5,6,11,8,9]))),Ut=(0,N.lazy)(()=>$(()=>import(`./WorkNotesDrawer-DwaUPJ_M.js`).then(e=>({default:e.WorkNotesDrawer})),__vite__mapDeps([16,1,2,3,5,6,11,8,9])));function Wt(){let e=Ot(X.themeMode);return e===`light`||e===`dark`||e===`system`?e:`system`}function Gt(e){return Object.fromEntries(e.map(e=>[e.id,e]))}function Kt(){return window.prompt(`Title for new feature`)?.trim()||null}function qt(e){let t=e.trim().replace(/[/\\]+$/,``);if(!t)return``;let n=t.split(/[/\\]/).filter(Boolean);return n[n.length-1]??t}function Jt(){let[e,t]=(0,N.useState)(null),[n,r]=(0,N.useState)(``),[i,a]=(0,N.useState)(null),[o,s]=(0,N.useState)(null),[c,l]=(0,N.useState)(null),[u,d]=(0,N.useState)([]),[f,p]=(0,N.useState)(null),[m,h]=(0,N.useState)({}),[v,C]=(0,N.useState)(!1),[w,E]=(0,N.useState)(null),[ie,oe]=(0,N.useState)(null),D=(0,N.useRef)({}),O=(0,N.useRef)(null),A=(0,N.useRef)(0),j=(0,N.useRef)({}),M=(0,N.useRef)(null),[P,ve]=(0,N.useState)(0),[ye,Ce]=(0,N.useState)(!1),[we,Te]=(0,N.useState)(!1),[F,Ee]=(0,N.useState)(!1),[De,Oe]=(0,N.useState)(!1),[ke,Ae]=(0,N.useState)(!1),[I,Ne]=(0,N.useState)(null),[Pe,Le]=(0,N.useState)(!1),[Re,Be]=(0,N.useState)(null),[Ve,He]=(0,N.useState)(()=>{let e=Ot(X.refreshSec);if(e!==null){let t=parseInt(e,10);if(!Number.isNaN(t)&&t>=0&&t<=120)return t}return 5}),Ue=(0,N.useRef)(null),[R,z]=(0,N.useState)(()=>{let e=Ot(X.splitPct);if(e){let t=Number(e);if(!Number.isNaN(t)&&t>=15&&t<=85)return t}return 42}),[We,B]=(0,N.useState)(()=>Ot(X.showInheritedDeps)===`1`),[Ge,qe]=(0,N.useState)(()=>Ot(X.highlightDepChain)!==`0`),[V,Je]=(0,N.useState)(!1),[Ye,H]=(0,N.useState)(Wt);(0,N.useLayoutEffect)(()=>{if(!i)return;let e=Et(X.themeMode,i);(e===`light`||e===`dark`||e===`system`)&&H(e);let t=Et(X.refreshSec,i);if(t!==null){let e=parseInt(t,10);!Number.isNaN(e)&&e>=0&&e<=120&&He(e)}let n=Et(X.splitPct,i);if(n){let e=Number(n);!Number.isNaN(e)&&e>=15&&e<=85&&z(e)}B(Et(X.showInheritedDeps,i)===`1`),qe(Et(X.highlightDepChain,i)!==`0`)},[i]);let[U,Xe]=(0,N.useState)(null),[W,Ze]=(0,N.useState)(new Set),Qe=(0,N.useRef)(null),$e=(0,N.useRef)(null),et=(0,N.useRef)(null),[tt,nt]=(0,N.useState)(52),rt=(0,N.useRef)(!1),it=(0,N.useRef)(!1),at=(0,N.useRef)(new Set),ot=(0,N.useRef)(null);(0,N.useEffect)(()=>{at.current=W},[W]),(0,N.useEffect)(()=>{ot.current=U},[U]),(0,N.useEffect)(()=>{O.current=f},[f]),(0,N.useEffect)(()=>{A.current=P},[P]),(0,N.useLayoutEffect)(()=>{document.documentElement.setAttribute(`data-theme`,Ye)},[Ye]),(0,N.useEffect)(()=>{Dt(X.themeMode,i,Ye)},[Ye,i]),(0,N.useLayoutEffect)(()=>{let e=M.current;if(!e)return;let t=()=>ve(e.getBoundingClientRect().bottom);t();let n=new ResizeObserver(t);return n.observe(e),()=>n.disconnect()},[]);let lt=(0,N.useCallback)(()=>{let e=$e.current;if(!e)return;let t=e.querySelector(`thead`),n=e.querySelector(`tbody tr.outline-gap-tr`),r=t?.getBoundingClientRect().height??0,i=n?.getBoundingClientRect().height,a=r+(typeof i==`number`&&i>0?i:8);a>0&&nt(Math.round(a*10)/10)},[]);(0,N.useLayoutEffect)(()=>{if(!e)return;let t=$e.current;if(!t)return;let n=new ResizeObserver(()=>{lt()});n.observe(t);let r=t.querySelector(`thead`),i=t.querySelector(`tbody`);r&&n.observe(r),i&&n.observe(i);let a=requestAnimationFrame(()=>{lt()});return()=>{cancelAnimationFrame(a),n.disconnect()}},[e,lt]);let ut=(0,N.useCallback)(async()=>{try{let e=await b();Ne({vision:e.vision_needs_completion,constitution:e.constitution_needs_completion})}catch{Ne(null)}},[]),dt=(0,N.useCallback)(async()=>{try{Be(await x())}catch{Be(null)}},[]),G=(0,N.useCallback)(async()=>{s(null);try{let[e,n]=await Promise.all([S(),fetch(`/api/repo`).then(e=>e.json())]);t(e),Ue.current=e.fingerprint;let i=n;r(i.repo_root||``),a(typeof i.repo_id==`string`?i.repo_id:null),l(t=>t&&e.ordered_ids.includes(t)?t:e.ordered_ids[0]??null);try{Ue.current=await ne()}catch{}ut(),dt()}catch(e){s(String(e))}},[ut,dt]),ft=(0,N.useCallback)(async e=>{try{await _(e)}catch(e){throw e instanceof y&&await G(),e}await G()},[G]),{busy:K,busyLabel:pt,runRoadmapAction:mt}=At(),q=It(),ht=(0,N.useRef)(new Map),[gt,_t]=(0,N.useState)(0),vt=(0,N.useCallback)(()=>_t(e=>e+1|0),[]),yt=(0,N.useCallback)((e,t)=>mt(e,async()=>{if(await jt(t,G)===`conflict`){s(`Roadmap or workspace changed elsewhere; the view was refreshed. Retry if needed.`);return}await G()}),[mt,G]),J=(0,N.useMemo)(()=>!e||ht.current.size===0?e:st(e,[...ht.current.values()]),[e,gt]),Y=(0,N.useCallback)((e,t,n,r,i)=>{let a=Ft();q.begin(a,n,r),t&&(ht.current.set(a,t),vt());let o=e=>{ht.current.delete(a)&&vt(),e?q.end(a):q.fail(a)};return mt(e,async()=>{try{if(await jt(i,G)===`conflict`){s(`Roadmap or workspace changed elsewhere; the view was refreshed. Retry if needed.`),o(!1);return}await G(),o(!0)}catch(e){throw o(!1),e}})},[mt,G,q,vt]),bt=(0,N.useMemo)(()=>({onConcurrencyConflict:G}),[G]),xt=(0,N.useCallback)(e=>Y(`Updating outline…`,{kind:`indent`,nodeId:e},[e],`indent`,async()=>{await ee(e)}).catch(e=>s(String(e))),[Y]),Ct=(0,N.useCallback)(e=>Y(`Updating outline…`,{kind:`outdent`,nodeId:e},[e],`outdent`,async()=>{await re(e)}).catch(e=>s(String(e))),[Y]);(0,N.useEffect)(()=>{g(()=>Ue.current)},[]),(0,N.useEffect)(()=>{G()},[G]),(0,N.useEffect)(()=>{let e=window.setInterval(()=>{dt()},12e3);return()=>window.clearInterval(e)},[dt]),(0,N.useEffect)(()=>{Dt(X.splitPct,i,String(R))},[R,i]),(0,N.useEffect)(()=>{Dt(X.showInheritedDeps,i,We?`1`:`0`)},[We,i]),(0,N.useEffect)(()=>{Dt(X.highlightDepChain,i,Ge?`1`:`0`)},[Ge,i]),(0,N.useEffect)(()=>{Dt(X.refreshSec,i,String(Ve))},[Ve,i]),(0,N.useEffect)(()=>{if(Ve<=0)return;let e=window.setInterval(()=>{(async()=>{try{let e=await ne(),t=Ue.current;t!==null&&e!==t?await mt(`Checking for updates…`,()=>G()):Ue.current=e}catch{}})()},Ve*1e3);return()=>window.clearInterval(e)},[Ve,mt,G]);let Tt=(0,N.useCallback)(()=>{Xe(null),Ze(new Set)},[]),kt=(0,N.useCallback)(()=>{Ze(new Set)},[]),Nt=(0,N.useCallback)(async()=>{let e=ot.current;if(!e)return;let t=[...at.current],n=[...t].sort().join(` `);try{s(null),await Y(`Saving dependencies…`,{kind:`dep`,nodeId:e,explicitNodeKeys:t},[e],`dep`,async()=>{await ae(e,[{key:`dependencies`,value:n}]),Tt()})}catch(e){s(String(e))}},[Y,Tt]),Pt=(0,N.useMemo)(()=>qt(n),[n]),Z=(0,N.useMemo)(()=>J?Gt(J.nodes):{},[J]),Lt=(0,N.useMemo)(()=>{if(!c)return!0;let e=Z[c]?.parent_id;return e==null||e===``},[c,Z]),Q=(0,N.useMemo)(()=>J?.ordered_ids?Me(J.ordered_ids,Z,J.registry_by_node):{},[J,Z]),$=(0,N.useMemo)(()=>J?.ordered_ids?je(J.ordered_ids,Z,J.registry_by_node):{},[J,Z]),Jt=(0,N.useMemo)(()=>J?.ordered_ids?V?J.ordered_ids.filter(e=>($[e]??``).trim().toLowerCase()!==`complete`):J.ordered_ids:[],[J?.ordered_ids,V,$]),Yt=(0,N.useMemo)(()=>{if(!J?.ordered_ids||!J.row_depths)return[];if(!V)return J.row_depths;let e=[];for(let t=0;t<J.ordered_ids.length;t++)($[J.ordered_ids[t]]??``).trim().toLowerCase()!==`complete`&&e.push(J.row_depths[t]??0);return e},[J?.ordered_ids,J?.row_depths,V,$]),Xt=(0,N.useMemo)(()=>!V||!J?.dependency_depths?0:Ie(Jt,J.dependency_depths),[V,J?.dependency_depths,Jt]);(0,N.useEffect)(()=>{if(!V||!J?.ordered_ids?.length)return;let e=e=>($[e]??``).trim().toLowerCase()===`complete`;l(t=>t&&!e(t)?t:J.ordered_ids.find(t=>!e(t))??null)},[V,J?.ordered_ids,$]),(0,N.useEffect)(()=>{!V||!U||($[U]??``).trim().toLowerCase()===`complete`&&Tt()},[V,U,$,Tt]);let Zt=(0,N.useMemo)(()=>{if(!J?.ordered_ids)return{};let e=J.registry_by_node??{},t=J.git_enrichment??{},n={};for(let r of J.ordered_ids)n[r]=k(Z[r],e[r],t[r],Q[r]);return n},[J,Z,Q]),Qt=(0,N.useMemo)(()=>{if(!J?.ordered_ids)return{};let e=J.registry_by_node??{},t=J.git_workflow?.resolved?.git_branch_current??null,n={};for(let r of J.ordered_ids)n[r]=ce(r,e,t);return n},[e]),$t=(0,N.useMemo)(()=>{if(!c)return!1;let e=(Z[c]?.status)?.trim()||`Not Started`,t=Q[c]??e,n=Zt[c]??t;return se(!!Qt[c],t,n)},[c,Z,Q,Zt,Qt]),en=(0,N.useMemo)(()=>J?.nodes?Object.fromEntries(J.nodes.map(e=>[e.node_key,e.id])):{},[e]),tn=(0,N.useMemo)(()=>!Ge||!c?null:Fe(c,Z,en),[Ge,c,Z,en]),nn=(0,N.useCallback)(e=>{let t=J?.nodes.find(t=>t.id===e);t&&(Xe(e),Ze(new Set(t.dependencies??[])))},[J?.nodes]),rn=(0,N.useCallback)(e=>{if(U===e){Nt();return}nn(e)},[U,Nt,nn]),an=(0,N.useCallback)(e=>{if(!U)return;let t=Z[U]?.node_key,n=Z[e]?.node_key;!t||!n||n===t||Ze(e=>{let t=new Set(e);return t.has(n)?t.delete(n):t.add(n),t})},[U,Z]),on=(0,N.useCallback)((e,t)=>{D.current[e]=t,h(t=>{if(!(e in t))return t;let n={...t};return delete n[e],n})},[]),sn=(0,N.useCallback)(e=>{d(t=>{let n=!t.includes(e),r=t.includes(e)?[...t.filter(t=>t!==e),e]:[...t,e];if(n){let n=O.current??t[t.length-1]??null,r=be(n?D.current[n]:void 0,A.current);h(t=>({...t,[e]:r}))}return r}),p(e)},[]),cn=(0,N.useCallback)(e=>{p(e),d(t=>t.includes(e)?[...t.filter(t=>t!==e),e]:t)},[]),ln=(0,N.useCallback)(e=>{d(t=>t.filter(t=>t!==e)),delete D.current[e],h(t=>{if(!(e in t))return t;let n={...t};return delete n[e],n})},[]),un=(0,N.useCallback)(()=>{!J||u.length===0||(v?(oe({...j.current}),C(!1),E(null),requestAnimationFrame(()=>{requestAnimationFrame(()=>oe(null))})):(j.current={...D.current},E(Se(xe(u,Z,J.ordered_ids),P)),C(!0)))},[J,u,Z,v,P]);(0,N.useEffect)(()=>{f!=null&&!u.includes(f)&&p(u.length?u[u.length-1]:null)},[u,f]),(0,N.useEffect)(()=>{u.length===0&&v&&(C(!1),E(null))},[u.length,v]),(0,N.useEffect)(()=>{!v||!J||u.length===0||E(Se(xe(u,Z,J.ordered_ids),P))},[v,J,u,Z,P]);let dn=!c||J?.outline_actions!=null&&c!=null&&J.outline_actions[c]?.can_indent===!1,fn=!c||J?.outline_actions!=null&&c!=null&&J.outline_actions[c]?.can_outdent===!1;(0,N.useEffect)(()=>{let e=e=>{if(u.length>0||ye||we||F)return;let t=e.target;if(t instanceof HTMLElement){if(t.isContentEditable)return;let e=t.tagName;if(e===`INPUT`||e===`TEXTAREA`||e===`SELECT`)return}if(e.key===`Escape`&&U){e.preventDefault(),Tt();return}if(e.key===`Enter`&&U){if(t instanceof HTMLElement){let e=t.tagName;if(e===`BUTTON`||e===`TEXTAREA`||e===`A`)return}e.preventDefault(),Nt();return}if(e.key===`Tab`&&c&&!U){if(t instanceof HTMLElement&&t.closest(`.app-header`))return;e.preventDefault(),e.shiftKey?fn||Ct(c):dn||xt(c)}};return window.addEventListener(`keydown`,e),()=>window.removeEventListener(`keydown`,e)},[u.length,ye,we,F,U,c,dn,fn,xt,Ct,Tt,Nt]);let pn=e=>{let t=$e.current,n=et.current;!t||!n||rt.current||(rt.current=!0,e===`left`?n.scrollTop=t.scrollTop:t.scrollTop=n.scrollTop,requestAnimationFrame(()=>{rt.current=!1}))},mn=e=>{e.preventDefault(),it.current=!0;let t=Qe.current;if(!t)return;e.target.setPointerCapture(e.pointerId);let n=e=>{if(!it.current)return;let n=t.getBoundingClientRect(),r=(e.clientX-n.left)/n.width*100;z(Math.min(85,Math.max(15,r)))},r=t=>{it.current=!1,e.target.releasePointerCapture(t.pointerId),window.removeEventListener(`pointermove`,n),window.removeEventListener(`pointerup`,r)};window.addEventListener(`pointermove`,n),window.addEventListener(`pointerup`,r)},hn=(0,N.useCallback)((e,t,n,r,i)=>{let a=Z[e]?.parent_id??null,o=ct({token:Ft(),title:n,type:r,parentId:a});Y(i,{kind:`add`,placeholder:o,referenceNodeId:e,position:t},[o.id],`add`,async()=>{await te(e,t,n,r)}).catch(e=>s(String(e)))},[Z,Y]),gn=()=>{if(!c)return;let e=Kt();e&&hn(c,`above`,e,`task`,`Adding task…`)},_n=()=>{if(!c)return;let e=Kt();e&&hn(c,`below`,e,`task`,`Adding task…`)},vn=()=>{if(!c)return;let e=Kt();e&&hn(c,`above`,e,`gate`,`Inserting gate…`)},yn=e=>{let t=Kt();t&&hn(e,`above`,t,`task`,`Adding task…`)},bn=()=>{if(!c)return;let e=Z[c]?.id??c;window.confirm(`Remove task "${e}" from the roadmap? This cannot be undone.`)&&(s(null),Y(`Removing task…`,{kind:`delete`,nodeId:c},[c],`delete`,async()=>{await T(c)}).catch(e=>s(String(e))))},xn=!!(Re?.scope_dirty&&!Re?.blocked&&!K);return(0,L.jsx)(_e,{value:bt,children:(0,L.jsx)(Mt,{api:q,children:(0,L.jsxs)(`div`,{className:`app-shell`,children:[K?(0,L.jsxs)(`div`,{className:`roadmap-sync-banner`,role:`status`,"aria-live":`polite`,"aria-label":pt??`Working`,children:[(0,L.jsx)(`span`,{className:`roadmap-sync-spinner`,"aria-hidden":`true`}),(0,L.jsx)(`span`,{children:pt??`Working…`})]}):null,(0,L.jsxs)(`div`,{className:`app-chrome`,"aria-busy":K,inert:K?!0:void 0,children:[(0,L.jsxs)(`header`,{ref:M,className:`app-header`,children:[(0,L.jsxs)(`div`,{className:`app-header-row1`,children:[(0,L.jsx)(`div`,{className:`app-header-row1-left`,children:(0,L.jsxs)(`h1`,{className:`app-title-line`,children:[(0,L.jsx)(`span`,{className:`app-title-core`,children:`specy-road — PM Gantt`}),Pt?(0,L.jsxs)(L.Fragment,{children:[(0,L.jsxs)(`span`,{className:`repo-root-sep`,"aria-hidden":`true`,children:[` `,`-`,` `]}),(0,L.jsx)(`span`,{className:`repo-root-folder`,title:n,"aria-label":`Repository root: ${n}`,children:Pt})]}):null]})}),(0,L.jsxs)(`div`,{className:`app-header-row1-actions`,children:[(0,L.jsx)(ze,{gitWorkflow:J?.git_workflow,integrationBranchAutoFf:J?.integration_branch_auto_ff,registryRemoteOverlay:J?.registry_overlay?.enabled===!0}),(0,L.jsx)(`button`,{type:`button`,className:`app-header-icon-btn app-header-tile-btn`,"aria-pressed":V,disabled:K,title:V?`Show all roadmap rows, including completed work`:`Hide rows whose subtree is effectively complete`,"aria-label":V?`Show completed work`:`Hide completed work`,onClick:()=>Je(e=>!e),children:V?`Show Complete`:`Hide Complete`}),u.length>0?(0,L.jsx)(`button`,{type:`button`,className:`app-header-icon-btn app-header-tile-btn`,"aria-pressed":v,disabled:K,title:v?`Restore task dialogs to their positions before tiling`:`Tile open task dialogs by dependency, left to right`,"aria-label":v?`Untile task dialogs`:`Tile task dialogs`,onClick:()=>un(),children:`Tile`}):null,(0,L.jsx)(`button`,{type:`button`,className:`app-header-icon-btn`,"aria-label":`Settings`,title:`Settings`,disabled:K,onClick:()=>Ce(!0),children:(0,L.jsx)(ge,{})})]})]}),(0,L.jsx)(`div`,{className:`app-header-row2`,children:(0,L.jsxs)(`div`,{className:`app-header-row2-inner`,children:[(0,L.jsxs)(`div`,{className:`app-header-toolbar`,children:[(0,L.jsx)(`button`,{type:`button`,className:xn?`toolbar-icon-btn toolbar-icon-btn--publish-ready`:`toolbar-icon-btn`,disabled:K,title:Re?.blocked?`Publish blocked — see dialog for details`:xn?`Publish roadmap changes (ready to share)`:`Publish roadmap changes to the remote repository`,"aria-label":`Publish roadmap changes`,onClick:()=>{Le(!0),dt()},children:(0,L.jsx)(he,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||!c||$t,title:$t?`Editing is disabled while this task is in active development (in progress, MR state, or checkout matches the registered branch)`:`Edit selected task`,"aria-label":`Edit selected task`,onClick:()=>{c&&sn(c)},children:(0,L.jsx)(me,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||dn,title:`Indent`,"aria-label":`Indent`,onClick:()=>{c&&xt(c)},children:(0,L.jsx)(pe,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||fn,title:`Outdent`,"aria-label":`Outdent`,onClick:()=>{c&&Ct(c)},children:(0,L.jsx)(le,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||!c,title:`Add task above selection`,"aria-label":`Add task above selection`,onClick:gn,children:(0,L.jsx)(ue,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||!c,title:`Add task below selection`,"aria-label":`Add task below selection`,onClick:_n,children:(0,L.jsx)(fe,{})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn`,disabled:K||!c||Lt,title:c?Lt?`Gate rows must sit under a vision or phase — select an indented row`:`Insert gate at selection (above this row); display ids renumber. PM hold, not dev pickup`:`Select a row first`,"aria-label":`Insert gate at selection`,onClick:vn,children:(0,L.jsx)(`span`,{className:`toolbar-gate-icon`,"aria-hidden":`true`,children:`G`})}),(0,L.jsx)(`button`,{type:`button`,className:`toolbar-icon-btn toolbar-icon-btn-danger`,disabled:K||!c,title:`Delete selected row`,"aria-label":`Delete selected row`,onClick:bn,children:(0,L.jsx)(de,{})})]}),(0,L.jsxs)(`div`,{className:`app-header-docs-row`,children:[(0,L.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,L.jsxs)(`div`,{className:I?.vision?`app-header-doc-tooltip app-header-doc-tooltip--incomplete`:`app-header-doc-tooltip`,children:[(0,L.jsx)(`button`,{type:`button`,className:I?.vision?`app-header-doc-btn app-header-doc-btn--incomplete`:`app-header-doc-btn`,"aria-describedby":I?.vision?`gov-tip-vision`:void 0,onClick:()=>Ee(!0),children:`Vision`}),I?.vision?(0,L.jsxs)(`div`,{id:`gov-tip-vision`,role:`tooltip`,className:`app-header-doc-tip`,children:[`Needs completion — replace the starter text in`,` `,(0,L.jsx)(`code`,{children:`vision.md`}),` with your product vision.`]}):null]})}),(0,L.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,L.jsxs)(`div`,{className:I?.constitution?`app-header-doc-tooltip app-header-doc-tooltip--incomplete`:`app-header-doc-tooltip`,children:[(0,L.jsx)(`button`,{type:`button`,className:I?.constitution?`app-header-doc-btn app-header-doc-btn--incomplete`:`app-header-doc-btn`,"aria-describedby":I?.constitution?`gov-tip-constitution`:void 0,onClick:()=>Te(!0),children:`Constitution`}),I?.constitution?(0,L.jsxs)(`div`,{id:`gov-tip-constitution`,role:`tooltip`,className:`app-header-doc-tip`,children:[`Needs completion — customize`,` `,(0,L.jsx)(`code`,{children:`purpose.md`}),` and `,(0,L.jsx)(`code`,{children:`principles.md`}),` `,`beyond the scaffold templates.`]}):null]})}),(0,L.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,L.jsx)(`button`,{type:`button`,className:`app-header-doc-btn`,onClick:()=>Oe(!0),children:`Shared docs`})}),(0,L.jsx)(`div`,{className:`app-header-doc-slot`,children:(0,L.jsx)(`button`,{type:`button`,className:`app-header-doc-btn`,onClick:()=>Ae(!0),children:`Session notes`})})]})]})})]}),o?(0,L.jsx)(`p`,{style:{padding:`0 0.75rem`,color:`crimson`},children:o}):null,J?(0,L.jsxs)(`div`,{className:`split`,ref:Qe,children:[(0,L.jsx)(`div`,{className:`outline-wrap`,ref:$e,style:{flex:`0 0 ${R}%`,width:`${R}%`},onScroll:()=>pn(`left`),children:(0,L.jsx)(St,{orderedIds:Jt,nodesById:Z,displayStatusById:Q,outlineStatusById:Zt,rowDepths:Yt,reorderLocked:V,interactionLocked:K,selectedId:c,prHints:J.pr_hints,gitEnrichment:J.git_enrichment,dependencyInheritance:J.dependency_inheritance,registryByNode:J.registry_by_node,gitBranchCurrent:J.git_workflow?.resolved?.git_branch_current??null,gitUserName:J.git_workflow?.resolved?.git_user_name??null,depEditId:U,depDraftKeys:W,onToggleDepCandidate:an,onDepCellActivate:rn,onApplyDepEdit:()=>void Nt(),onCancelDepEdit:Tt,onClearDepDraft:kt,onSelect:l,onDoubleClick:e=>{l(e),sn(e)},performRoadmapMutation:yt,performOptimisticMutation:Y,onMutationError:e=>s(e),onGapInsert:yn})}),(0,L.jsx)(`div`,{className:`split-resize-handle`,role:`separator`,"aria-orientation":`vertical`,"aria-label":`Resize feature list and chart`,onPointerDown:mn}),(0,L.jsx)(`div`,{className:`gantt-wrap`,ref:et,onScroll:()=>pn(`right`),children:(0,L.jsx)(Ke,{orderedIds:Jt,nodesById:Z,displayStatusById:Q,gitCheckoutById:Qt,registryByNode:J.registry_by_node,gitEnrichment:J.git_enrichment,stackHeaderPx:tt,depthOffset:Xt,depths:J.dependency_depths,spans:J.dependency_spans,edges:J.edges,showInheritedEdges:We,selectedId:c,highlightRowIds:tn,onSelect:l,onChartBackgroundMouseDown:U?()=>void Nt():void 0})})]}):(0,L.jsx)(`p`,{style:{padding:`1rem`},children:`Loading roadmap…`}),(0,L.jsxs)(N.Suspense,{fallback:null,children:[J?u.map((e,t)=>{let n=Z[e];if(!n)return null;let r=n.status?.trim()||`Not Started`,i=Q[e]??r,a=Zt[e]??i,o=se(!!Qt[e],i,a),c=u.length>1;return(0,L.jsx)(Rt,{node:n,pmDisplayStatusResolved:Q[e],allNodes:J.nodes,modalStorageKey:e,stackZIndex:50+t,backdropPassThrough:c,closeOnEscape:t===u.length-1,headerMinTop:P,spawnInitialRect:m[e],editTileMode:v,tileRect:w?.[e]??null,resumeFreeRect:ie?.[e]??null,titleBarActive:f===e,onActivate:()=>cn(e),onRectCommit:t=>on(e,t),dependencyInheritance:J.dependency_inheritance?.[e],registryByNode:J.registry_by_node,gitEnrichment:J.git_enrichment,prHints:J.pr_hints,onClose:()=>ln(e),onOpenNode:sn,onPersisted:()=>void mt(`Saving task…`,G).catch(e=>s(String(e))),readOnlyCheckout:o},e)}):null,(0,L.jsx)(zt,{open:we,onClose:()=>{Te(!1),ut()}}),(0,L.jsx)(Ht,{open:F,onClose:()=>{Ee(!1),ut()}}),(0,L.jsx)(Vt,{open:De,onClose:()=>Oe(!1)}),(0,L.jsx)(Ut,{open:ke,onClose:()=>Ae(!1)}),(0,L.jsx)(Bt,{open:ye,onClose:()=>Ce(!1),themeMode:Ye,onThemeModeChange:H,highlightDepChain:Ge,onHighlightDepChainChange:qe,showInheritedDeps:We,onShowInheritedDepsChange:B,refreshSec:Ve,onRefreshSecChange:He}),(0,L.jsx)(wt,{open:Pe,onClose:()=>Le(!1),status:Re,onRefreshStatus:dt,onPublish:ft,headerMinTop:P})]})]})]})})})}(0,P.createRoot)(document.getElementById(`root`)).render((0,L.jsx)(N.StrictMode,{children:(0,L.jsx)(Jt,{})}));
